Build a Keras Multilayer Perceptron

Utility function to build a Keras MLP.

build_keras_net(
  n_in,
  n_out,
  nodes = c(32, 32),
  layer_pars = list(),
  activation = "linear",
  act_pars = list(),
  dropout = 0.1,
  batch_norm = TRUE,
  batch_pars = list()
)

Arguments

n_in
(integer(1))
Number of input features.

n_out
(integer(1))
Number of targets.

nodes
(numeric())
Hidden nodes in network, each element in vector represents number of hidden nodes in respective layer.

layer_pars
(list())
Arguments passed to keras::layer_dense.

activation
(character(1))
Activation function passed to keras::layer_activation. Default is linear.

act_pars
(list())
Parameters for activation function, see keras::layer_activation.

dropout
(numeric(1))
Optional dropout layer, if NULL then no dropout layer added otherwise either same dropout will be added to all layers.

batch_norm
(logical(1))
If TRUE (default) then batch normalisation is applied to all layers.

batch_pars
(list())
Parameters for batch normalisation, see keras::layer_batch_normalization.

Details

This function is a helper for R users with less Python experience. Currently it is limited to simple MLPs and with identical layers. More advanced networks will require manual creation with keras.

Examples

build_keras_net(10, 1)

#> Model
#> Model: "sequential"
#> ________________________________________________________________________________
#> Layer (type)                        Output Shape                    Param #     
#> ================================================================================
#> dense (Dense)                       (None, 32)                      352         
#> ________________________________________________________________________________
#> activation (Activation)             (None, 32)                      0           
#> ________________________________________________________________________________
#> batch_normalization (BatchNormaliza (None, 32)                      128         
#> ________________________________________________________________________________
#> dropout (Dropout)                   (None, 32)                      0           
#> ________________________________________________________________________________
#> dense_1 (Dense)                     (None, 32)                      1056        
#> ________________________________________________________________________________
#> activation_1 (Activation)           (None, 32)                      0           
#> ________________________________________________________________________________
#> batch_normalization_1 (BatchNormali (None, 32)                      128         
#> ________________________________________________________________________________
#> dropout_1 (Dropout)                 (None, 32)                      0           
#> ________________________________________________________________________________
#> dense_2 (Dense)                     (None, 1)                       33          
#> ________________________________________________________________________________
#> activation_2 (Activation)           (None, 1)                       0           
#> ________________________________________________________________________________
#> batch_normalization_2 (BatchNormali (None, 1)                       4           
#> ________________________________________________________________________________
#> dropout_2 (Dropout)                 (None, 1)                       0           
#> ================================================================================
#> Total params: 1,701
#> Trainable params: 1,571
#> Non-trainable params: 130
#> ________________________________________________________________________________
#> 
#> 

build_keras_net(n_in = 10, n_out = 1, nodes = c(4, 4, 4), activation = "elu")

#> Model
#> Model: "sequential_1"
#> ________________________________________________________________________________
#> Layer (type)                        Output Shape                    Param #     
#> ================================================================================
#> dense_3 (Dense)                     (None, 4)                       44          
#> ________________________________________________________________________________
#> activation_3 (Activation)           (None, 4)                       0           
#> ________________________________________________________________________________
#> batch_normalization_3 (BatchNormali (None, 4)                       16          
#> ________________________________________________________________________________
#> dropout_3 (Dropout)                 (None, 4)                       0           
#> ________________________________________________________________________________
#> dense_4 (Dense)                     (None, 4)                       20          
#> ________________________________________________________________________________
#> activation_4 (Activation)           (None, 4)                       0           
#> ________________________________________________________________________________
#> batch_normalization_4 (BatchNormali (None, 4)                       16          
#> ________________________________________________________________________________
#> dropout_4 (Dropout)                 (None, 4)                       0           
#> ________________________________________________________________________________
#> dense_5 (Dense)                     (None, 4)                       20          
#> ________________________________________________________________________________
#> activation_5 (Activation)           (None, 4)                       0           
#> ________________________________________________________________________________
#> batch_normalization_5 (BatchNormali (None, 4)                       16          
#> ________________________________________________________________________________
#> dropout_5 (Dropout)                 (None, 4)                       0           
#> ________________________________________________________________________________
#> dense_6 (Dense)                     (None, 1)                       5           
#> ________________________________________________________________________________
#> activation_6 (Activation)           (None, 1)                       0           
#> ________________________________________________________________________________
#> batch_normalization_6 (BatchNormali (None, 1)                       4           
#> ________________________________________________________________________________
#> dropout_6 (Dropout)                 (None, 1)                       0           
#> ================================================================================
#> Total params: 141
#> Trainable params: 115
#> Non-trainable params: 26
#> ________________________________________________________________________________
#> 
#>