Build a Pytorch Multilayer Perceptron

Utility function to build an MLP with a choice of activation function and weight initialization with optional dropout and batch normalization.

build_pytorch_net(
  n_in,
  n_out,
  nodes = c(32, 32),
  activation = "relu",
  act_pars = list(),
  dropout = 0.1,
  bias = TRUE,
  batch_norm = TRUE,
  batch_pars = list(eps = 1e-05, momentum = 0.1, affine = TRUE),
  init = "uniform",
  init_pars = list()
)

Arguments

n_in
(integer(1))
Number of input features.

n_out
(integer(1))
Number of targets.

nodes
(numeric())
Hidden nodes in network, each element in vector represents number of hidden nodes in respective layer.

activation
(character(1)|list())
Activation function, can either be a single character and the same function is used in all layers, or a list of length length(nodes). See get_pycox_activation for options.

act_pars
(list())
Passed to get_pycox_activation.

dropout
(numeric())
Optional dropout layer, if NULL then no dropout layer added otherwise either a single numeric which will be added to all layers or a vector of differing drop-out amounts.

bias
(logical(1))
If TRUE (default) then a bias parameter is added to all linear layers.

batch_norm
(logical(1))
If TRUE (default) then batch normalisation is applied to all layers.

batch_pars
(list())
Parameters for batch normalisation, see reticulate::py_help(torch$nn$BatchNorm1d).

init
(character(1))
Weight initialization method. See get_pycox_init for options.

init_pars
(list())
Passed to get_pycox_init.

Details

This function is a helper for R users with less Python experience. Currently it is limited to simple MLPs. More advanced networks will require manual creation with reticulate.

Examples

build_pytorch_net(10, 1)

#> Sequential(
#>   (L0): Linear(in_features=10, out_features=32, bias=True)
#>   (A0): ReLU()
#>   (BN0): BatchNorm1d(32, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
#>   (D0): Dropout(p=0.1, inplace=False)
#>   (L1): Linear(in_features=32, out_features=32, bias=True)
#>   (A1): ReLU()
#>   (BN1): BatchNorm1d(32, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
#>   (D1): Dropout(p=0.1, inplace=False)
#>   (2): Linear(in_features=32, out_features=1, bias=True)
#> )

build_pytorch_net(
  n_in = 10, n_out = 1, nodes = c(4, 4, 4), activation = "elu",
  act_pars = list(alpha = 0.5), dropout = c(0.2, 0.1, 0.6),
  batch_norm = TRUE, init = "kaiming_normal", init_pars = list(non_linearity = "relu"))

#> Sequential(
#>   (L0): Linear(in_features=10, out_features=4, bias=True)
#>   (A0): ELU(alpha=0.5)
#>   (BN0): BatchNorm1d(4, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
#>   (D0): Dropout(p=0.2, inplace=False)
#>   (L1): Linear(in_features=4, out_features=4, bias=True)
#>   (A1): ELU(alpha=0.5)
#>   (BN1): BatchNorm1d(4, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
#>   (D1): Dropout(p=0.2, inplace=False)
#>   (L2): Linear(in_features=4, out_features=4, bias=True)
#>   (A2): ELU(alpha=0.5)
#>   (BN2): BatchNorm1d(4, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
#>   (D2): Dropout(p=0.1, inplace=False)
#>   (3): Linear(in_features=4, out_features=1, bias=True)
#> )