Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      0  27   1 5.690      1
#> 2      0  28   1 5.544      1
#> 3      1  46   0 4.902      1
#> 4      0  21   1 5.668      1
#> 5      0  37   1 5.322      1
#> 6      1  42   1 6.932      0
#> 7      0  49   0 3.480      1
#> 8      1  32   1 6.932      0
#> 9      1  25   1 6.790      1
#> 10     0  26   0 3.537      1
#> 11     0  29   1 5.599      1
#> 12     0  26   1 5.483      1
#> 13     0  43   1 5.534      1
#> 14     0  49   1 5.461      1
#> 15     1  38   1 6.757      1
#> 16     0  33   1 5.635      1
#> 17     1  48   1 6.929      1
#> 18     1  25   1 6.932      0
#> 19     1  46   0 5.002      1
#> 20     0  47   1 5.487      1
#> 21     1  27   1 6.932      0
#> 22     0  35   1 5.353      1
#> 23     1  34   0 4.926      1
#> 24     1  31   1 6.932      0
#> 25     0  25   1 5.506      1
#> 26     1  27   1 6.932      0
#> 27     0  42   1 5.538      1
#> 28     0  32   1 5.525      1
#> 29     1  46   1 6.932      0
#> 30     1  23   1 6.932      1
#> 31     1  22   1 6.932      0
#> 32     1  39   1 6.932      0
#> 33     1  31   1 6.932      0
#> 34     0  40   1 5.407      1
#> 35     1  26   1 6.932      0
#> 36     0  27   1 5.517      1
#> 37     1  39   1 6.932      0
#> 38     1  45   1 6.714      1
#> 39     1  35   1 6.857      1
#> 40     1  47   1 6.932      0
#> 41     1  30   1 6.932      0
#> 42     1  25   1 6.932      0
#> 43     1  41   1 6.782      1
#> 44     1  37   1 6.746      1
#> 45     1  30   1 6.932      0
#> 46     1  40   1 6.932      0
#> 47     1  25   1 6.932      0
#> 48     0  25   1 5.433      1
#> 49     1  43   1 6.932      0
#> 50     1  50   1 6.889      1
#> 51     0  24   1 5.578      1
#> 52     1  44   0 4.966      1
#> 53     1  33   0 4.979      1
#> 54     1  29   0 5.090      1
#> 55     0  46   1 5.493      1
#> 56     1  44   1 6.932      0
#> 57     0  32   1 5.435      1
#> 58     1  46   1 6.912      1
#> 59     1  47   1 6.932      0
#> 60     0  41   1 5.368      1
#> 61     1  37   1 6.932      0
#> 62     1  39   0 4.938      1
#> 63     0  27   1 5.631      1
#> 64     1  28   1 6.932      0
#> 65     1  46   1 6.915      1
#> 66     0  41   1 5.544      1
#> 67     1  21   0 4.890      1
#> 68     0  42   1 5.470      1
#> 69     1  43   0 5.045      1
#> 70     1  23   1 6.932      0
#> 71     1  44   0 5.061      1
#> 72     0  21   1 5.456      1
#> 73     1  49   0 4.983      1
#> 74     1  39   0 5.017      1
#> 75     1  42   1 6.932      0
#> 76     0  34   0 3.414      1
#> 77     0  47   1 5.410      1
#> 78     0  29   1 5.592      1
#> 79     0  31   0 3.532      1
#> 80     0  26   1 5.391      1
#> 81     0  28   0 3.487      1
#> 82     1  36   1 6.932      0
#> 83     1  30   0 4.976      1
#> 84     1  23   1 6.884      1
#> 85     1  36   1 6.932      0
#> 86     0  21   1 5.153      1
#> 87     1  29   1 6.932      0
#> 88     1  46   1 6.932      0
#> 89     0  31   0 3.465      1
#> 90     0  48   1 5.536      1
#> 91     1  30   1 6.740      1
#> 92     0  45   1 5.515      1
#> 93     1  28   1 6.845      1
#> 94     0  48   1 5.545      1
#> 95     0  37   1 5.560      1
#> 96     1  22   0 4.836      1
#> 97     0  33   1 5.519      1
#> 98     0  43   1 5.394      1
#> 99     1  22   1 6.932      0
#> 100    0  37   1 5.406      1