Build a Keras Multilayer Perceptron

Utility function to build a Keras MLP.

build_keras_net(
  n_in,
  n_out,
  nodes = c(32L, 32L),
  layer_pars = list(),
  activation = "linear",
  act_pars = list(),
  dropout = 0.1,
  batch_norm = TRUE,
  batch_pars = list()
)

Arguments

n_in
(integer(1))
Number of input features.

n_out
(integer(1))
Number of targets.

nodes
(numeric())
Hidden nodes in network, each element in vector represents number of hidden nodes in respective layer.

layer_pars
(list())
Arguments passed to keras::layer_dense.

activation
(character(1))
Activation function passed to keras::layer_activation. Default is linear.

act_pars
(list())
Parameters for activation function, see keras::layer_activation.

dropout
(numeric(1))
Optional dropout layer, if NULL then no dropout layer added otherwise either same dropout will be added to all layers.

batch_norm
(logical(1))
If TRUE (default) then batch normalisation is applied to all layers.

batch_pars
(list())
Parameters for batch normalisation, see keras::layer_batch_normalization.

Details

This function is a helper for R users with less Python experience. Currently it is limited to simple MLPs and with identical layers. More advanced networks will require manual creation with keras.

Examples

# \donttest{
if (requireNamespaces("keras")) {
  build_keras_net(4L, 2L)

  build_keras_net(n_in = 4L, n_out = 2L, nodes = c(32L, 64L, 32L),
    activation = "elu", dropout = 0.4)
}

#> Model
#> Model: "sequential_1"
#> ________________________________________________________________________________
#> Layer (type)                        Output Shape                    Param #     
#> ================================================================================
#> dense_3 (Dense)                     (None, 32)                      160         
#> ________________________________________________________________________________
#> activation_3 (Activation)           (None, 32)                      0           
#> ________________________________________________________________________________
#> batch_normalization_3 (BatchNormali (None, 32)                      128         
#> ________________________________________________________________________________
#> dropout_3 (Dropout)                 (None, 32)                      0           
#> ________________________________________________________________________________
#> dense_4 (Dense)                     (None, 64)                      2112        
#> ________________________________________________________________________________
#> activation_4 (Activation)           (None, 64)                      0           
#> ________________________________________________________________________________
#> batch_normalization_4 (BatchNormali (None, 64)                      256         
#> ________________________________________________________________________________
#> dropout_4 (Dropout)                 (None, 64)                      0           
#> ________________________________________________________________________________
#> dense_5 (Dense)                     (None, 32)                      2080        
#> ________________________________________________________________________________
#> activation_5 (Activation)           (None, 32)                      0           
#> ________________________________________________________________________________
#> batch_normalization_5 (BatchNormali (None, 32)                      128         
#> ________________________________________________________________________________
#> dropout_5 (Dropout)                 (None, 32)                      0           
#> ________________________________________________________________________________
#> dense_6 (Dense)                     (None, 2)                       66          
#> ________________________________________________________________________________
#> activation_6 (Activation)           (None, 2)                       0           
#> ________________________________________________________________________________
#> batch_normalization_6 (BatchNormali (None, 2)                       8           
#> ________________________________________________________________________________
#> dropout_6 (Dropout)                 (None, 2)                       0           
#> ================================================================================
#> Total params: 4,938
#> Trainable params: 4,678
#> Non-trainable params: 260
#> ________________________________________________________________________________
#> 
#> 

# }