Build a Pytorch Multilayer Perceptron

Utility function to build an MLP with a choice of activation function and weight initialization with optional dropout and batch normalization.

build_pytorch_net(
  n_in,
  n_out,
  nodes = c(32, 32),
  activation = "relu",
  act_pars = list(),
  dropout = 0.1,
  bias = TRUE,
  batch_norm = TRUE,
  batch_pars = list(eps = 1e-05, momentum = 0.1, affine = TRUE),
  init = "uniform",
  init_pars = list()
)

Arguments

n_in
(integer(1))
Number of input features.

n_out
(integer(1))
Number of targets.

nodes
(numeric())
Hidden nodes in network, each element in vector represents number of hidden nodes in respective layer.

activation
(character(1)|list())
Activation function, can either be a single character and the same function is used in all layers, or a list of length length(nodes). See get_pycox_activation for options.

act_pars
(list())
Passed to get_pycox_activation.

dropout
(numeric())
Optional dropout layer, if NULL then no dropout layer added otherwise either a single numeric which will be added to all layers or a vector of differing drop-out amounts.

bias
(logical(1))
If TRUE (default) then a bias parameter is added to all linear layers.

batch_norm
(logical(1))
If TRUE (default) then batch normalisation is applied to all layers.

batch_pars
(list())
Parameters for batch normalisation, see reticulate::py_help(torch$nn$BatchNorm1d).

init
(character(1))
Weight initialization method. See get_pycox_init for options.

init_pars
(list())
Passed to get_pycox_init.

Details

This function is a helper for R users with less Python experience. Currently it is limited to simple MLPs. More advanced networks will require manual creation with reticulate.

Examples

# \donttest{
if (requireNamespaces("reticulate")) {
  build_pytorch_net(4L, 2L, nodes = c(32, 64, 32), activation = "selu")

  # pass parameters to activation and initializer functions
  build_pytorch_net(4L, 2L, activation = "elu", act_pars = list(alpha = 0.1),
  init  = "kaiming_uniform", init_pars = list(mode = "fan_out"))
}

#> Sequential(
#>   (L0): Linear(in_features=4, out_features=32, bias=True)
#>   (A0): ELU(alpha=0.1)
#>   (BN0): BatchNorm1d(32, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
#>   (D0): Dropout(p=0.1, inplace=False)
#>   (L1): Linear(in_features=32, out_features=32, bias=True)
#>   (A1): ELU(alpha=0.1)
#>   (BN1): BatchNorm1d(32, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
#>   (D1): Dropout(p=0.1, inplace=False)
#>   (2): Linear(in_features=32, out_features=2, bias=True)
#> )

# }