Predict Method for DNNSurv

Predicted values from a fitted object of class dnnsurv.

# S3 method for dnnsurv
predict(
  object,
  newdata,
  batch_size = 32L,
  verbose = 0L,
  steps = NULL,
  callbacks = NULL,
  type = c("survival", "risk", "all"),
  distr6 = FALSE,
  ...
)

Arguments

object
(dnnsurv(1))
Object of class inheriting from "dnnsurv".

newdata
(data.frame(1))
Testing data of data.frame like object, internally is coerced with stats::model.matrix(). If missing then training data from fitted object is used.

batch_size
(integer(1))
Passed to keras::predict.keras.engine.training.Model, elements in each batch.

verbose
(integer(1))
Level of verbosity for printing, 0 or 1.

steps
(integer(1))
Number of batches before evaluation finished, see keras::predict.keras.engine.training.Model.

callbacks
(list())
Optional callbacks to apply during prediction.

type
(numeric(1))
Type of predicted value. Choices are survival probabilities over all time-points in training data ("survival") or a relative risk ranking ("risk"), which is the mean cumulative hazard function over all time-points, or both ("all").

distr6
(logical(1))
If FALSE (default) and type is "survival" or "all" returns data.frame of survival probabilities, otherwise returns a distr6::VectorDistribution().

...
ANY
Currently ignored.