Predict Method for DNNSurv

Predicted values from a fitted object of class dnnsurv.

# S3 method for dnnsurv
predict(
  object,
  newdata,
  batch_size = 32L,
  verbose = 0L,
  steps = NULL,
  callbacks = NULL,
  type = c("survival", "risk", "all"),
  distr6 = FALSE,
  ...
)

Arguments

object
(dnnsurv(1))
Object of class inheriting from "dnnsurv".

newdata
(data.frame(1))
Testing data of data.frame like object, internally is coerced with stats::model.matrix(). If missing then training data from fitted object is used.

batch_size
(integer(1))
Passed to keras::predict.keras.engine.training.Model, elements in each batch.

verbose
(integer(1))
Level of verbosity for printing, 0 or 1.

steps
(integer(1))
Number of batches before evaluation finished, see keras::predict.keras.engine.training.Model.

callbacks
(list())
Optional callbacks to apply during prediction.

type
(character(1))
Type of predicted value. Choices are survival probabilities over all time-points in training data ("survival") or a relative risk ranking ("risk"), which is the negative mean survival time so higher rank implies higher risk of event, or both ("all").

distr6
(logical(1))
If FALSE (default) and type is "survival" or "all" returns matrix of survival probabilities, otherwise returns a distr6::VectorDistribution().

...
ANY
Currently ignored.

Value

A numeric if type = "risk", a distr6::VectorDistribution() (if distr6 = TRUE) and type = "survival"; a matrix if (distr6 = FALSE) and type = "survival" where entries are survival probabilities with rows of observations and columns are time-points; or a list combining above if type = "all".

Examples

# \donttest{
if (requireNamespaces(c("keras", "pseudo")))
  fit <- dnnsurv(data = simsurvdata(10))

  # predict survival matrix and relative risks
  predict(fit, simsurvdata(10), type = "all")

#> $surv
#>           3.545   4.39125    5.2375   6.08375      6.93
#>  [1,] 0.9271273 0.8595788 0.7969550 0.7388736 0.6850496
#>  [2,] 0.9271237 0.8595737 0.7969485 0.7388639 0.6850401
#>  [3,] 0.9271738 0.8596706 0.7970849 0.7390274 0.6852320
#>  [4,] 0.9271173 0.8595632 0.7969343 0.7388452 0.6850199
#>  [5,] 0.9271230 0.8595681 0.7969387 0.7388561 0.6850269
#>  [6,] 0.9271617 0.8596483 0.7970549 0.7389904 0.6851897
#>  [7,] 0.9271495 0.8596232 0.7970193 0.7389493 0.6851403
#>  [8,] 0.9271469 0.8596180 0.7970118 0.7389406 0.6851299
#>  [9,] 0.9271600 0.8596445 0.7970496 0.7389846 0.6851825
#> [10,] 0.9271223 0.8595666 0.7969364 0.7388535 0.6850236
#> 
#> $risk
#>  [1] -1.609301 -1.609346 -1.608395 -1.609444 -1.609417 -1.608602 -1.608848
#>  [8] -1.608900 -1.608637 -1.609434
#> 

  # return as distribution
  if (requireNamespaces("distr6")) {
    predict(fit, simsurvdata(10), distr6 = TRUE)
  }

#> WeightDisc1 WeightDisc2 ... WeightDisc9 WeightDisc10 

# }