Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      1  22   0 4.994      1
#> 2      1  22   0 5.122      1
#> 3      1  37   0 4.924      1
#> 4      0  20   1 5.555      1
#> 5      1  32   1 6.877      0
#> 6      0  25   1 5.358      1
#> 7      0  33   1 5.488      1
#> 8      0  32   0 3.542      1
#> 9      0  20   1 5.337      1
#> 10     1  34   1 6.875      1
#> 11     1  45   1 6.877      0
#> 12     0  26   1 5.387      1
#> 13     1  24   0 4.949      1
#> 14     1  32   1 6.877      0
#> 15     0  21   1 5.287      1
#> 16     0  32   1 5.448      1
#> 17     1  41   1 6.877      0
#> 18     1  28   0 5.013      1
#> 19     0  31   1 5.462      1
#> 20     1  36   1 6.877      0
#> 21     0  36   1 5.402      1
#> 22     0  34   1 5.420      1
#> 23     0  43   1 5.346      1
#> 24     0  26   0 3.541      1
#> 25     0  20   0 3.491      1
#> 26     1  42   1 6.877      0
#> 27     1  29   1 6.830      1
#> 28     1  42   0 4.954      1
#> 29     0  33   0 3.550      1
#> 30     0  28   0 3.553      1
#> 31     0  20   1 5.346      1
#> 32     1  28   1 6.877      0
#> 33     1  36   0 4.922      1
#> 34     1  46   1 6.877      0
#> 35     1  24   0 5.065      1
#> 36     0  42   1 5.636      1
#> 37     1  40   1 6.877      0
#> 38     1  24   1 6.877      0
#> 39     1  28   1 6.877      0
#> 40     1  38   1 6.877      0
#> 41     1  37   1 6.877      0
#> 42     0  43   0 3.474      1
#> 43     0  38   1 5.357      1
#> 44     1  43   1 6.877      0
#> 45     1  43   1 6.877      0
#> 46     1  27   0 4.808      1
#> 47     0  24   1 5.448      1
#> 48     1  20   0 4.690      1
#> 49     0  29   1 5.197      1
#> 50     1  27   0 4.984      1
#> 51     1  27   0 4.951      1
#> 52     0  36   0 3.543      1
#> 53     1  47   0 5.089      1
#> 54     0  32   1 5.491      1
#> 55     1  21   1 6.801      1
#> 56     0  30   1 5.412      1
#> 57     1  48   1 6.877      0
#> 58     0  20   1 5.347      1
#> 59     1  36   1 6.877      0
#> 60     1  39   0 4.976      1
#> 61     0  32   0 3.501      1
#> 62     1  22   1 6.877      0
#> 63     1  35   1 6.877      0
#> 64     0  28   1 5.379      1
#> 65     0  38   1 5.511      1
#> 66     1  20   1 6.511      1
#> 67     0  42   0 3.530      1
#> 68     1  30   1 6.877      0
#> 69     0  20   0 3.586      1
#> 70     0  30   1 5.476      1
#> 71     1  49   1 6.743      1
#> 72     0  43   1 5.363      1
#> 73     1  35   1 6.870      1
#> 74     1  41   0 4.986      1
#> 75     0  42   1 5.493      1
#> 76     1  21   1 6.877      0
#> 77     1  32   1 6.877      0
#> 78     1  49   1 6.877      0
#> 79     0  32   1 5.487      1
#> 80     0  20   1 5.590      1
#> 81     0  40   1 5.438      1
#> 82     0  49   0 3.478      1
#> 83     0  47   1 5.530      1
#> 84     0  24   1 5.250      1
#> 85     1  46   1 6.705      1
#> 86     0  37   0 3.433      1
#> 87     0  39   1 5.591      1
#> 88     0  39   1 5.366      1
#> 89     1  45   1 6.877      0
#> 90     0  42   0 3.495      1
#> 91     1  39   1 6.877      0
#> 92     1  22   1 6.877      0
#> 93     1  37   1 6.877      0
#> 94     1  35   0 4.979      1
#> 95     0  49   1 5.529      1
#> 96     1  48   0 4.857      1
#> 97     1  44   1 6.877      0
#> 98     1  45   1 6.876      1
#> 99     1  35   1 6.877      0
#> 100    1  44   1 6.877      0