Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cutoff = 20)

Arguments

n
(integer(1))
Number of samples

trt, age, sex
(numeric(1))
Coefficients for covariates.

cutoff
(numeric(1))
Cutoff for determining Type I censoring.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()

#>     sexF age trt   time status
#> 1      0  42   1 16.412      1
#> 2      0  33   0  1.813      1
#> 3      1  46   1 20.000      0
#> 4      0  35   1 17.380      1
#> 5      1  25   1 20.000      0
#> 6      1  44   1 20.000      0
#> 7      1  43   0 13.759      1
#> 8      1  26   1 20.000      0
#> 9      1  30   1 20.000      0
#> 10     0  27   1 14.508      1
#> 11     1  24   0 13.279      1
#> 12     1  40   1 20.000      0
#> 13     1  38   1 20.000      0
#> 14     0  26   1 17.108      1
#> 15     0  33   0  1.896      1
#> 16     1  26   1 20.000      0
#> 17     0  21   1 16.791      1
#> 18     0  34   1 17.153      1
#> 19     1  49   0 13.363      1
#> 20     0  23   0  1.000      1
#> 21     0  26   1 17.561      1
#> 22     1  27   0 12.900      1
#> 23     0  39   1 16.165      1
#> 24     0  28   1 17.602      1
#> 25     1  30   1 20.000      0
#> 26     0  40   0  2.604      1
#> 27     0  27   1 16.527      1
#> 28     1  45   1 20.000      0
#> 29     0  40   1 16.529      1
#> 30     0  36   1 17.866      1
#> 31     1  24   1 20.000      0
#> 32     1  29   0 13.040      1
#> 33     1  47   1 20.000      0
#> 34     0  25   1 17.625      1
#> 35     0  47   1 17.905      1
#> 36     1  37   1 20.000      0
#> 37     0  32   1 17.758      1
#> 38     0  36   1 16.525      1
#> 39     0  30   1 16.437      1
#> 40     0  47   0  2.583      1
#> 41     0  46   0  2.397      1
#> 42     1  23   0 14.116      1
#> 43     0  42   1 16.561      1
#> 44     1  37   0 12.660      1
#> 45     0  35   1 16.558      1
#> 46     1  29   1 20.000      0
#> 47     1  50   1 20.000      0
#> 48     1  48   0 13.888      1
#> 49     1  50   0 13.843      1
#> 50     1  46   1 20.000      0
#> 51     0  47   1 16.952      1
#> 52     1  34   1 20.000      0
#> 53     0  23   1 17.584      1
#> 54     1  28   1 20.000      0
#> 55     1  40   1 20.000      0
#> 56     0  31   0  1.779      1
#> 57     1  45   1 20.000      0
#> 58     0  40   1 17.166      1
#> 59     1  30   1 20.000      0
#> 60     1  24   1 20.000      0
#> 61     0  48   1 16.172      1
#> 62     0  38   1 17.174      1
#> 63     1  42   0 13.076      1
#> 64     1  29   0 14.638      1
#> 65     0  40   1 17.015      1
#> 66     0  28   1 16.837      1
#> 67     0  27   1 17.459      1
#> 68     1  33   1 20.000      0
#> 69     0  45   0  2.233      1
#> 70     0  25   1 14.459      1
#> 71     0  37   1 16.283      1
#> 72     0  39   0  2.477      1
#> 73     0  27   0  2.536      1
#> 74     0  40   1 17.615      1
#> 75     0  48   1 16.815      1
#> 76     1  23   1 20.000      0
#> 77     1  45   0 12.901      1
#> 78     0  30   0  1.927      1
#> 79     1  45   0 13.699      1
#> 80     0  44   0  2.231      1
#> 81     1  41   1 20.000      0
#> 82     0  24   1 18.345      1
#> 83     0  25   1 17.365      1
#> 84     1  28   1 20.000      0
#> 85     1  46   1 20.000      0
#> 86     1  44   1 20.000      0
#> 87     1  29   0 13.563      1
#> 88     1  23   1 20.000      0
#> 89     0  32   0  2.571      1
#> 90     1  21   1 20.000      0
#> 91     0  31   1 16.564      1
#> 92     1  29   0 13.878      1
#> 93     1  36   0 13.466      1
#> 94     1  20   0 13.436      1
#> 95     0  32   0  1.784      1
#> 96     1  24   0 12.694      1
#> 97     1  23   1 20.000      0
#> 98     1  46   1 20.000      0
#> 99     0  35   1 17.684      1
#> 100    0  46   0  2.796      1