Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      0  30   1 5.473      1
#> 2      1  20   1 6.502      1
#> 3      1  33   1 6.835      0
#> 4      1  40   0 5.016      1
#> 5      0  25   1 5.656      1
#> 6      0  32   0 3.586      1
#> 7      0  38   1 5.301      1
#> 8      1  47   1 6.835      0
#> 9      1  46   1 6.835      0
#> 10     0  42   1 5.405      1
#> 11     0  39   1 5.427      1
#> 12     0  38   1 5.523      1
#> 13     0  46   1 5.500      1
#> 14     0  46   1 5.478      1
#> 15     1  41   1 6.835      0
#> 16     0  31   1 5.562      1
#> 17     1  44   1 6.835      0
#> 18     0  46   0 3.501      1
#> 19     0  39   1 5.564      1
#> 20     1  21   1 6.835      0
#> 21     1  30   1 6.835      0
#> 22     0  23   1 5.575      1
#> 23     0  39   1 5.171      1
#> 24     1  22   0 4.842      1
#> 25     0  47   1 5.104      1
#> 26     1  24   1 6.747      1
#> 27     1  21   1 6.835      0
#> 28     0  32   1 5.213      1
#> 29     0  49   1 5.401      1
#> 30     1  26   1 6.835      0
#> 31     0  23   0 3.506      1
#> 32     1  28   1 6.835      0
#> 33     0  32   1 5.479      1
#> 34     0  21   0 3.437      1
#> 35     1  21   0 4.981      1
#> 36     0  43   0 3.453      1
#> 37     0  40   0 3.500      1
#> 38     0  47   1 5.443      1
#> 39     1  37   1 6.835      0
#> 40     1  22   0 4.701      1
#> 41     0  33   0 3.578      1
#> 42     0  49   1 5.380      1
#> 43     1  25   1 6.835      0
#> 44     1  49   1 6.835      0
#> 45     0  38   1 5.542      1
#> 46     0  27   0 3.488      1
#> 47     1  40   1 6.835      0
#> 48     1  44   1 6.835      0
#> 49     1  47   1 6.835      0
#> 50     1  26   1 6.835      0
#> 51     1  38   1 6.835      0
#> 52     1  32   0 5.080      1
#> 53     1  38   0 5.005      1
#> 54     0  25   1 5.361      1
#> 55     1  43   1 6.835      0
#> 56     1  34   0 4.925      1
#> 57     0  30   0 3.483      1
#> 58     0  47   1 5.491      1
#> 59     0  20   0 3.439      1
#> 60     0  41   1 5.464      1
#> 61     1  44   1 6.835      0
#> 62     1  25   0 4.923      1
#> 63     0  36   1 5.511      1
#> 64     0  40   0 3.484      1
#> 65     0  48   1 5.450      1
#> 66     1  32   1 6.835      0
#> 67     1  44   1 6.823      1
#> 68     1  41   1 6.835      0
#> 69     0  40   0 3.493      1
#> 70     0  40   1 5.392      1
#> 71     1  26   0 4.961      1
#> 72     0  23   1 5.491      1
#> 73     0  32   1 5.236      1
#> 74     1  46   0 5.062      1
#> 75     0  31   1 5.489      1
#> 76     0  29   1 5.508      1
#> 77     1  36   0 4.981      1
#> 78     1  23   1 6.835      0
#> 79     1  24   1 6.835      0
#> 80     0  42   1 5.426      1
#> 81     1  37   1 6.835      0
#> 82     0  27   1 5.684      1
#> 83     1  39   1 6.835      0
#> 84     1  42   0 4.834      1
#> 85     0  46   0 3.487      1
#> 86     0  46   0 3.499      1
#> 87     1  50   1 6.835      0
#> 88     1  44   0 4.933      1
#> 89     0  29   1 5.527      1
#> 90     0  27   1 5.217      1
#> 91     0  26   1 5.589      1
#> 92     0  41   1 5.444      1
#> 93     0  38   0 3.316      1
#> 94     1  24   1 6.835      0
#> 95     0  44   1 5.459      1
#> 96     1  47   1 6.835      0
#> 97     1  47   0 4.939      1
#> 98     1  37   1 6.835      0
#> 99     1  43   0 4.913      1
#> 100    0  32   1 5.353      1