Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      0  29   0 3.405      1
#> 2      0  41   1 5.576      1
#> 3      1  36   0 5.004      1
#> 4      0  39   1 5.415      1
#> 5      1  35   0 4.957      1
#> 6      1  47   1 6.847      0
#> 7      0  26   1 5.535      1
#> 8      0  43   0 3.508      1
#> 9      0  34   0 3.524      1
#> 10     1  43   1 6.847      0
#> 11     1  39   1 6.847      0
#> 12     0  27   1 5.459      1
#> 13     1  28   1 6.708      1
#> 14     0  21   0 3.385      1
#> 15     1  33   1 6.847      0
#> 16     0  21   0 3.373      1
#> 17     0  44   0 3.488      1
#> 18     0  26   1 5.090      1
#> 19     1  43   1 6.847      0
#> 20     0  39   1 5.449      1
#> 21     0  39   0 3.462      1
#> 22     0  43   0 3.472      1
#> 23     1  22   1 6.847      0
#> 24     1  43   1 6.847      0
#> 25     0  49   0 3.400      1
#> 26     1  26   0 4.457      1
#> 27     0  36   1 5.460      1
#> 28     1  34   1 6.847      0
#> 29     1  44   1 6.847      0
#> 30     0  21   1 5.334      1
#> 31     0  49   1 5.476      1
#> 32     1  40   1 6.847      0
#> 33     0  22   0 3.509      1
#> 34     1  31   1 6.737      1
#> 35     1  32   0 5.072      1
#> 36     1  22   0 5.026      1
#> 37     0  43   0 3.518      1
#> 38     0  41   1 5.529      1
#> 39     1  45   1 6.847      0
#> 40     1  38   1 6.847      0
#> 41     0  42   0 3.440      1
#> 42     0  32   1 5.310      1
#> 43     0  45   1 5.347      1
#> 44     0  24   0 3.565      1
#> 45     1  50   1 6.847      0
#> 46     0  39   1 5.418      1
#> 47     1  22   1 6.847      0
#> 48     0  41   0 3.507      1
#> 49     0  30   1 5.469      1
#> 50     0  31   1 5.489      1
#> 51     0  29   0 2.916      1
#> 52     1  34   0 4.989      1
#> 53     1  33   1 6.847      0
#> 54     0  37   1 5.415      1
#> 55     1  37   1 6.847      0
#> 56     0  50   1 5.514      1
#> 57     1  27   1 6.820      1
#> 58     1  41   1 6.847      0
#> 59     1  30   1 6.664      1
#> 60     0  44   1 5.243      1
#> 61     0  41   1 5.555      1
#> 62     0  27   1 5.471      1
#> 63     0  39   1 5.392      1
#> 64     1  27   1 6.847      0
#> 65     1  50   0 4.946      1
#> 66     0  29   1 5.424      1
#> 67     0  30   0 3.452      1
#> 68     0  48   0 3.496      1
#> 69     0  29   1 5.276      1
#> 70     1  43   1 6.847      0
#> 71     0  22   1 5.264      1
#> 72     1  36   0 4.874      1
#> 73     1  24   1 6.847      0
#> 74     1  36   0 4.816      1
#> 75     0  41   1 5.440      1
#> 76     0  30   1 5.523      1
#> 77     1  30   1 6.847      0
#> 78     1  25   1 6.847      0
#> 79     0  26   1 5.490      1
#> 80     0  24   0 3.537      1
#> 81     1  21   1 6.847      0
#> 82     1  24   0 4.981      1
#> 83     1  41   1 6.847      0
#> 84     1  45   0 5.004      1
#> 85     1  45   1 6.847      0
#> 86     0  27   1 5.078      1
#> 87     0  27   0 3.370      1
#> 88     1  22   1 6.805      1
#> 89     0  43   1 5.389      1
#> 90     1  28   1 6.847      0
#> 91     1  22   1 6.845      1
#> 92     0  31   1 5.465      1
#> 93     1  44   1 6.847      0
#> 94     1  50   0 5.006      1
#> 95     0  42   1 5.369      1
#> 96     1  22   1 6.847      0
#> 97     1  24   1 6.847      0
#> 98     0  50   1 5.483      1
#> 99     1  32   1 6.847      0
#> 100    1  22   1 6.824      1