Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cutoff = NULL, cens = 0.3)

Arguments

n
(integer(1))
Number of samples

trt, age, sex
(numeric(1))
Coefficients for covariates.

cutoff
(numeric(1))
Deprecated, in future use cens.

cens
(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()

#>     sexF age trt  time status
#> 1      0  42   1 5.365      1
#> 2      0  33   0 3.382      1
#> 3      1  46   1 6.814      0
#> 4      0  35   1 5.496      1
#> 5      1  25   1 6.814      0
#> 6      1  44   1 6.814      0
#> 7      1  43   0 5.004      1
#> 8      1  26   1 6.814      0
#> 9      1  30   1 6.814      0
#> 10     0  27   1 5.106      1
#> 11     1  24   0 4.939      1
#> 12     1  40   1 6.814      0
#> 13     1  38   1 6.814      0
#> 14     0  26   1 5.459      1
#> 15     0  33   0 3.393      1
#> 16     1  26   1 6.787      1
#> 17     0  21   1 5.416      1
#> 18     0  34   1 5.465      1
#> 19     1  49   0 4.951      1
#> 20     0  23   0 3.271      1
#> 21     0  26   1 5.521      1
#> 22     1  27   0 4.888      1
#> 23     0  39   1 5.331      1
#> 24     0  28   1 5.526      1
#> 25     1  30   1 6.814      0
#> 26     0  40   0 3.489      1
#> 27     0  27   1 5.380      1
#> 28     1  45   1 6.814      0
#> 29     0  40   1 5.381      1
#> 30     0  36   1 5.562      1
#> 31     1  24   1 6.814      0
#> 32     1  29   0 4.907      1
#> 33     1  47   1 6.814      0
#> 34     0  25   1 5.530      1
#> 35     0  47   1 5.568      1
#> 36     1  37   1 6.814      0
#> 37     0  32   1 5.548      1
#> 38     0  36   1 5.380      1
#> 39     0  30   1 5.368      1
#> 40     0  47   0 3.486      1
#> 41     0  46   0 3.461      1
#> 42     1  23   0 5.053      1
#> 43     0  42   1 5.385      1
#> 44     1  37   0 4.855      1
#> 45     0  35   1 5.385      1
#> 46     1  29   1 6.814      0
#> 47     1  50   1 6.814      0
#> 48     1  48   0 5.022      1
#> 49     1  50   0 5.016      1
#> 50     1  46   1 6.814      0
#> 51     0  47   1 5.438      1
#> 52     1  34   1 6.814      0
#> 53     0  23   1 5.524      1
#> 54     1  28   1 6.814      0
#> 55     1  40   1 6.814      0
#> 56     0  31   0 3.377      1
#> 57     1  45   1 6.814      0
#> 58     0  40   1 5.467      1
#> 59     1  30   1 6.814      0
#> 60     1  24   1 6.814      0
#> 61     0  48   1 5.332      1
#> 62     0  38   1 5.468      1
#> 63     1  42   0 4.912      1
#> 64     1  29   0 5.124      1
#> 65     0  40   1 5.447      1
#> 66     0  28   1 5.422      1
#> 67     0  27   1 5.507      1
#> 68     1  33   1 6.694      1
#> 69     0  45   0 3.439      1
#> 70     0  25   1 5.099      1
#> 71     0  37   1 5.347      1
#> 72     0  39   0 3.472      1
#> 73     0  27   0 3.480      1
#> 74     0  40   1 5.528      1
#> 75     0  48   1 5.420      1
#> 76     1  23   1 6.814      0
#> 77     1  45   0 4.888      1
#> 78     0  30   0 3.397      1
#> 79     1  45   0 4.996      1
#> 80     0  44   0 3.438      1
#> 81     1  41   1 6.814      0
#> 82     0  24   1 5.627      1
#> 83     0  25   1 5.494      1
#> 84     1  28   1 6.814      0
#> 85     1  46   1 6.814      0
#> 86     1  44   1 6.814      0
#> 87     1  29   0 4.978      1
#> 88     1  23   1 6.814      0
#> 89     0  32   0 3.485      1
#> 90     1  21   1 6.814      0
#> 91     0  31   1 5.385      1
#> 92     1  29   0 5.021      1
#> 93     1  36   0 4.964      1
#> 94     1  20   0 4.960      1
#> 95     0  32   0 3.378      1
#> 96     1  24   0 4.860      1
#> 97     1  23   1 6.814      0
#> 98     1  46   1 6.814      0
#> 99     0  35   1 5.538      1
#> 100    0  46   0 3.515      1