Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      1  31   0 5.061      1
#> 2      0  27   0 3.436      1
#> 3      1  41   1 6.785      0
#> 4      0  29   0 3.558      1
#> 5      1  23   0 4.983      1
#> 6      0  29   1 5.253      1
#> 7      1  44   1 6.785      0
#> 8      0  38   1 5.552      1
#> 9      1  40   1 6.683      1
#> 10     1  21   1 6.785      0
#> 11     1  45   0 4.968      1
#> 12     1  33   1 6.785      0
#> 13     0  21   0 3.611      1
#> 14     0  23   1 5.429      1
#> 15     1  34   0 4.833      1
#> 16     1  35   0 5.029      1
#> 17     0  45   0 3.539      1
#> 18     0  41   1 5.467      1
#> 19     1  31   1 6.781      1
#> 20     1  46   0 5.025      1
#> 21     1  42   0 4.943      1
#> 22     1  31   0 4.854      1
#> 23     1  39   0 5.064      1
#> 24     1  33   1 6.785      0
#> 25     0  22   1 5.556      1
#> 26     1  50   1 6.785      0
#> 27     1  35   1 6.785      0
#> 28     0  49   1 5.453      1
#> 29     1  33   1 6.785      0
#> 30     1  27   1 6.776      1
#> 31     1  26   1 6.535      1
#> 32     1  20   0 5.007      1
#> 33     0  35   1 5.572      1
#> 34     0  26   1 5.513      1
#> 35     1  30   0 4.884      1
#> 36     1  44   1 6.785      0
#> 37     0  41   1 5.404      1
#> 38     0  28   1 5.320      1
#> 39     1  50   1 6.785      0
#> 40     1  39   0 5.018      1
#> 41     1  23   1 6.553      1
#> 42     1  44   1 6.781      1
#> 43     1  48   1 6.785      0
#> 44     0  24   1 5.238      1
#> 45     0  45   0 3.515      1
#> 46     1  27   1 6.785      0
#> 47     1  45   1 6.785      0
#> 48     1  23   1 6.785      0
#> 49     0  38   1 5.459      1
#> 50     0  21   1 5.137      1
#> 51     1  50   0 5.033      1
#> 52     1  44   1 6.785      0
#> 53     1  33   0 4.972      1
#> 54     1  50   0 4.930      1
#> 55     0  30   1 5.596      1
#> 56     1  23   0 4.935      1
#> 57     0  42   1 5.413      1
#> 58     1  47   1 6.785      0
#> 59     1  40   1 6.785      0
#> 60     1  36   0 4.959      1
#> 61     0  44   1 5.466      1
#> 62     0  30   0 3.466      1
#> 63     1  31   1 6.785      0
#> 64     1  46   1 6.785      0
#> 65     1  42   1 6.785      0
#> 66     1  43   1 6.785      0
#> 67     0  43   0 3.511      1
#> 68     1  38   1 6.785      0
#> 69     0  47   0 3.466      1
#> 70     0  43   1 5.480      1
#> 71     1  45   0 4.900      1
#> 72     0  45   0 3.548      1
#> 73     0  33   1 5.363      1
#> 74     0  28   1 5.313      1
#> 75     0  23   1 5.672      1
#> 76     0  28   1 5.475      1
#> 77     1  22   1 6.785      0
#> 78     0  35   1 5.184      1
#> 79     1  23   0 4.876      1
#> 80     1  46   0 5.028      1
#> 81     1  48   0 5.050      1
#> 82     0  50   0 3.520      1
#> 83     1  29   0 5.040      1
#> 84     0  46   1 5.419      1
#> 85     0  48   0 3.474      1
#> 86     1  36   0 4.930      1
#> 87     1  30   0 4.986      1
#> 88     1  22   1 6.785      0
#> 89     1  27   0 4.903      1
#> 90     1  39   1 6.785      0
#> 91     0  28   1 5.467      1
#> 92     0  28   1 5.590      1
#> 93     1  20   1 6.785      0
#> 94     0  39   1 5.535      1
#> 95     1  30   1 6.785      0
#> 96     1  45   1 6.785      0
#> 97     1  30   1 6.785      0
#> 98     1  44   1 6.785      0
#> 99     0  36   1 5.536      1
#> 100    0  42   1 5.514      1