Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      0  29   1 5.453      1
#> 2      1  21   1 6.884      0
#> 3      1  38   1 6.873      1
#> 4      1  36   0 4.886      1
#> 5      1  27   1 6.879      1
#> 6      0  41   1 5.528      1
#> 7      1  30   0 4.976      1
#> 8      1  48   1 6.884      0
#> 9      1  47   1 6.884      0
#> 10     1  27   1 6.884      0
#> 11     1  26   1 6.884      0
#> 12     1  47   1 6.884      0
#> 13     0  32   1 5.184      1
#> 14     1  41   1 6.884      0
#> 15     1  44   0 4.998      1
#> 16     1  38   1 6.884      0
#> 17     1  31   1 6.884      0
#> 18     1  38   1 6.884      0
#> 19     0  48   1 5.507      1
#> 20     0  43   1 5.585      1
#> 21     1  45   0 5.039      1
#> 22     1  22   1 6.827      1
#> 23     0  22   1 5.491      1
#> 24     0  49   0 3.537      1
#> 25     0  46   1 5.474      1
#> 26     1  48   1 6.884      0
#> 27     0  42   1 5.469      1
#> 28     0  47   1 5.509      1
#> 29     1  25   0 5.008      1
#> 30     1  27   0 4.942      1
#> 31     0  38   1 5.215      1
#> 32     0  44   1 5.494      1
#> 33     0  48   1 5.496      1
#> 34     0  41   1 5.557      1
#> 35     1  33   1 6.884      0
#> 36     1  36   1 6.884      0
#> 37     1  21   0 5.158      1
#> 38     1  22   1 6.884      0
#> 39     1  29   1 6.884      0
#> 40     0  44   1 5.516      1
#> 41     0  27   0 3.554      1
#> 42     1  21   0 4.842      1
#> 43     0  25   0 3.206      1
#> 44     1  27   1 6.884      0
#> 45     1  38   0 4.905      1
#> 46     0  31   0 3.420      1
#> 47     0  34   0 3.534      1
#> 48     1  22   1 6.875      1
#> 49     1  22   1 6.884      0
#> 50     0  26   1 5.456      1
#> 51     1  40   1 6.884      0
#> 52     1  48   1 6.884      0
#> 53     1  39   1 6.860      1
#> 54     1  26   0 5.016      1
#> 55     1  24   0 4.978      1
#> 56     1  45   1 6.884      0
#> 57     1  23   1 6.884      0
#> 58     1  21   0 5.035      1
#> 59     0  38   1 5.568      1
#> 60     0  26   1 5.521      1
#> 61     0  45   1 5.504      1
#> 62     0  43   1 5.507      1
#> 63     0  25   1 5.587      1
#> 64     0  30   1 5.476      1
#> 65     1  42   1 6.652      1
#> 66     0  49   1 5.499      1
#> 67     0  46   1 5.556      1
#> 68     1  21   1 6.884      0
#> 69     1  22   0 5.045      1
#> 70     0  39   0 3.450      1
#> 71     1  27   1 6.884      0
#> 72     0  33   0 3.483      1
#> 73     1  45   1 6.884      0
#> 74     0  39   1 5.300      1
#> 75     0  47   1 5.404      1
#> 76     0  29   1 5.465      1
#> 77     0  21   0 3.472      1
#> 78     1  44   1 6.817      1
#> 79     1  47   1 6.884      0
#> 80     0  41   0 3.488      1
#> 81     0  21   1 5.351      1
#> 82     0  37   1 5.361      1
#> 83     0  28   1 5.475      1
#> 84     1  45   1 6.884      0
#> 85     0  45   0 3.396      1
#> 86     1  40   1 6.870      1
#> 87     0  33   1 5.505      1
#> 88     1  31   1 6.884      0
#> 89     0  48   1 5.496      1
#> 90     1  27   0 4.913      1
#> 91     0  24   0 3.599      1
#> 92     1  28   1 6.884      0
#> 93     0  30   0 3.434      1
#> 94     1  42   1 6.884      0
#> 95     1  42   1 6.884      0
#> 96     1  40   0 4.942      1
#> 97     1  45   0 5.030      1
#> 98     0  39   1 5.541      1
#> 99     0  21   1 5.485      1
#> 100    0  32   0 3.530      1