Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      0  40   1 5.553      1
#> 2      1  41   1 6.835      1
#> 3      1  47   1 6.920      0
#> 4      0  30   1 5.278      1
#> 5      1  29   1 6.920      0
#> 6      0  40   0 3.372      1
#> 7      1  49   0 4.982      1
#> 8      1  37   1 6.920      0
#> 9      0  25   0 3.506      1
#> 10     1  25   1 6.920      0
#> 11     1  48   1 6.920      0
#> 12     0  39   1 5.456      1
#> 13     0  42   1 5.520      1
#> 14     1  46   0 4.954      1
#> 15     1  35   1 6.918      1
#> 16     1  44   1 6.895      1
#> 17     0  30   0 3.431      1
#> 18     0  33   0 3.520      1
#> 19     1  22   1 6.890      1
#> 20     1  46   0 5.002      1
#> 21     1  23   1 6.920      0
#> 22     0  43   1 5.564      1
#> 23     1  34   0 4.719      1
#> 24     0  22   1 5.615      1
#> 25     1  28   0 4.823      1
#> 26     1  45   0 4.975      1
#> 27     1  37   1 6.920      0
#> 28     0  34   0 3.478      1
#> 29     1  47   1 6.780      1
#> 30     1  29   0 4.894      1
#> 31     0  22   1 5.585      1
#> 32     1  48   1 6.920      0
#> 33     1  41   1 6.920      0
#> 34     0  40   1 5.553      1
#> 35     0  22   1 5.387      1
#> 36     1  23   1 6.546      1
#> 37     1  24   1 6.920      0
#> 38     0  38   0 3.510      1
#> 39     0  22   1 5.352      1
#> 40     0  39   1 5.447      1
#> 41     1  22   1 6.920      0
#> 42     1  26   1 6.920      0
#> 43     0  26   1 5.414      1
#> 44     1  31   1 6.918      1
#> 45     1  44   1 6.826      1
#> 46     1  32   0 4.699      1
#> 47     0  48   0 3.467      1
#> 48     1  22   1 6.920      0
#> 49     1  27   1 6.487      1
#> 50     1  27   0 5.059      1
#> 51     1  48   1 6.920      0
#> 52     0  26   1 5.508      1
#> 53     0  38   1 5.413      1
#> 54     0  46   1 5.536      1
#> 55     0  26   0 3.596      1
#> 56     1  27   1 6.920      0
#> 57     1  31   1 6.920      0
#> 58     1  44   1 6.920      0
#> 59     1  34   1 6.920      0
#> 60     0  48   1 5.472      1
#> 61     1  45   1 6.920      0
#> 62     0  25   0 3.451      1
#> 63     0  38   1 5.481      1
#> 64     1  20   1 6.701      1
#> 65     0  39   1 5.445      1
#> 66     1  40   1 6.788      1
#> 67     1  28   1 6.920      0
#> 68     0  25   1 5.592      1
#> 69     0  21   1 5.456      1
#> 70     0  33   0 3.475      1
#> 71     0  34   0 3.377      1
#> 72     0  31   1 5.572      1
#> 73     0  23   1 5.484      1
#> 74     1  30   0 5.036      1
#> 75     1  24   0 4.806      1
#> 76     0  35   1 5.392      1
#> 77     0  44   0 3.536      1
#> 78     1  39   1 6.920      0
#> 79     1  29   1 6.920      0
#> 80     1  22   1 6.920      0
#> 81     1  33   1 6.920      0
#> 82     1  44   0 4.949      1
#> 83     1  49   1 6.833      1
#> 84     1  45   1 6.920      0
#> 85     0  37   1 5.546      1
#> 86     1  44   1 6.920      0
#> 87     0  40   1 5.467      1
#> 88     0  35   0 3.473      1
#> 89     1  45   0 4.952      1
#> 90     0  42   1 5.479      1
#> 91     1  42   1 6.920      0
#> 92     1  32   1 6.908      1
#> 93     1  45   1 6.873      1
#> 94     1  30   1 6.920      0
#> 95     0  39   1 5.409      1
#> 96     1  48   0 5.026      1
#> 97     1  21   1 6.697      1
#> 98     1  38   1 6.920      0
#> 99     0  36   0 3.471      1
#> 100    1  35   1 6.920      0