Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      0  22   0 3.502      1
#> 2      1  28   1 5.927      0
#> 3      1  39   1 5.927      0
#> 4      0  42   0 3.495      1
#> 5      1  38   1 5.927      0
#> 6      1  26   0 5.125      1
#> 7      1  37   1 5.927      0
#> 8      1  27   1 5.927      0
#> 9      1  45   1 5.927      0
#> 10     0  49   0 3.415      1
#> 11     1  32   0 4.875      1
#> 12     0  45   1 5.509      1
#> 13     0  33   0 3.539      1
#> 14     0  46   1 5.389      1
#> 15     1  50   0 4.916      1
#> 16     1  37   0 4.917      1
#> 17     0  28   1 5.377      1
#> 18     1  23   1 5.927      0
#> 19     1  49   1 5.927      0
#> 20     1  21   0 4.904      1
#> 21     0  32   1 5.534      1
#> 22     1  28   1 5.927      0
#> 23     1  23   0 4.805      1
#> 24     1  28   0 4.899      1
#> 25     0  47   0 3.495      1
#> 26     0  28   1 5.548      1
#> 27     1  39   1 5.927      0
#> 28     0  42   0 3.489      1
#> 29     0  37   1 5.505      1
#> 30     1  38   0 4.890      1
#> 31     0  47   1 5.502      1
#> 32     0  41   1 5.458      1
#> 33     1  32   1 5.927      0
#> 34     0  31   0 3.572      1
#> 35     1  39   0 5.008      1
#> 36     1  23   1 5.927      0
#> 37     0  20   1 5.478      1
#> 38     0  27   1 4.978      1
#> 39     1  39   1 5.927      0
#> 40     0  28   1 5.475      1
#> 41     1  41   1 5.927      0
#> 42     0  37   1 5.519      1
#> 43     1  25   0 5.000      1
#> 44     0  29   1 5.647      1
#> 45     1  49   0 4.939      1
#> 46     1  48   1 5.927      0
#> 47     0  44   0 3.473      1
#> 48     1  34   1 5.927      0
#> 49     1  44   0 4.972      1
#> 50     0  43   1 5.215      1
#> 51     0  50   1 5.501      1
#> 52     0  47   1 5.533      1
#> 53     0  26   1 5.440      1
#> 54     1  32   1 5.927      0
#> 55     1  49   1 5.927      0
#> 56     0  37   1 5.623      1
#> 57     0  49   0 3.496      1
#> 58     1  33   1 5.927      0
#> 59     0  33   1 5.378      1
#> 60     1  47   1 5.927      0
#> 61     0  38   1 5.498      1
#> 62     0  22   0 3.506      1
#> 63     0  33   1 5.528      1
#> 64     1  26   0 4.887      1
#> 65     0  49   1 5.453      1
#> 66     1  41   0 4.969      1
#> 67     1  46   1 5.927      0
#> 68     1  43   1 5.927      0
#> 69     0  30   1 5.459      1
#> 70     1  47   0 4.957      1
#> 71     0  46   0 3.490      1
#> 72     0  27   1 5.527      1
#> 73     0  20   1 5.597      1
#> 74     1  49   1 5.927      0
#> 75     0  50   1 5.454      1
#> 76     0  23   1 5.506      1
#> 77     0  48   0 3.516      1
#> 78     0  36   0 3.466      1
#> 79     1  22   1 5.927      0
#> 80     1  27   1 5.927      0
#> 81     0  41   1 5.458      1
#> 82     0  29   1 5.104      1
#> 83     1  44   0 5.037      1
#> 84     0  25   1 5.461      1
#> 85     1  25   1 5.927      0
#> 86     0  30   0 3.402      1
#> 87     0  34   0 3.478      1
#> 88     0  46   1 5.541      1
#> 89     0  34   1 5.532      1
#> 90     0  47   0 3.506      1
#> 91     1  47   1 5.927      0
#> 92     0  47   0 3.519      1
#> 93     0  23   1 5.539      1
#> 94     1  26   1 5.927      0
#> 95     0  24   0 3.401      1
#> 96     1  40   0 4.896      1
#> 97     0  46   1 5.558      1
#> 98     0  48   1 5.431      1
#> 99     1  45   1 5.927      0
#> 100    1  25   1 5.927      0