Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cutoff = NULL, cens = 0.3)

Arguments

n
(integer(1))
Number of samples

trt, age, sex
(numeric(1))
Coefficients for covariates.

cutoff
(numeric(1))
Deprecated, in future use cens.

cens
(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()

#>     sexF age trt  time status
#> 1      0  24   1 5.523      1
#> 2      1  23   0 4.924      1
#> 3      1  31   1 6.775      0
#> 4      1  34   1 6.775      0
#> 5      0  44   1 5.433      1
#> 6      0  35   1 5.444      1
#> 7      1  23   1 6.775      0
#> 8      1  48   1 6.775      0
#> 9      0  41   1 5.443      1
#> 10     1  35   1 6.775      0
#> 11     0  36   1 5.468      1
#> 12     0  28   0 3.442      1
#> 13     0  41   0 3.413      1
#> 14     1  49   1 6.775      0
#> 15     1  40   1 6.775      0
#> 16     0  44   1 5.489      1
#> 17     0  22   1 5.583      1
#> 18     1  47   1 6.775      0
#> 19     1  45   1 6.775      0
#> 20     1  42   0 5.058      1
#> 21     0  31   1 5.394      1
#> 22     1  36   1 6.775      0
#> 23     0  34   1 5.437      1
#> 24     1  23   0 5.075      1
#> 25     1  35   1 6.742      1
#> 26     1  33   0 4.936      1
#> 27     1  36   1 6.775      0
#> 28     1  39   1 6.775      0
#> 29     1  25   1 6.451      1
#> 30     0  29   1 5.576      1
#> 31     1  47   0 5.052      1
#> 32     1  38   1 6.775      0
#> 33     0  39   0 3.461      1
#> 34     0  28   1 5.530      1
#> 35     0  30   0 3.359      1
#> 36     1  30   1 6.775      0
#> 37     0  43   0 3.467      1
#> 38     0  20   1 5.565      1
#> 39     1  22   1 6.775      0
#> 40     1  28   1 6.775      0
#> 41     0  38   0 3.416      1
#> 42     1  29   1 6.775      0
#> 43     0  23   1 5.660      1
#> 44     1  43   0 4.915      1
#> 45     0  43   0 3.415      1
#> 46     0  32   1 5.443      1
#> 47     1  22   1 6.775      0
#> 48     1  43   0 4.857      1
#> 49     1  36   1 6.775      0
#> 50     1  28   0 4.983      1
#> 51     1  28   0 4.805      1
#> 52     0  36   1 5.399      1
#> 53     0  26   1 5.577      1
#> 54     1  39   1 6.775      0
#> 55     1  43   1 6.775      0
#> 56     0  28   0 3.475      1
#> 57     1  46   0 4.973      1
#> 58     1  46   1 6.775      0
#> 59     1  21   1 6.752      1
#> 60     1  31   1 6.677      1
#> 61     1  25   1 6.775      0
#> 62     1  48   0 4.979      1
#> 63     0  47   1 5.272      1
#> 64     0  26   1 5.504      1
#> 65     0  46   1 5.486      1
#> 66     1  20   1 6.775      0
#> 67     0  47   1 5.554      1
#> 68     1  32   1 6.775      0
#> 69     1  22   0 5.083      1
#> 70     0  39   0 3.447      1
#> 71     1  47   1 6.773      1
#> 72     0  45   0 3.465      1
#> 73     1  35   1 6.775      0
#> 74     1  38   1 6.775      0
#> 75     1  34   0 5.007      1
#> 76     0  47   1 5.455      1
#> 77     1  40   0 4.986      1
#> 78     1  32   1 6.771      1
#> 79     1  37   0 4.905      1
#> 80     0  27   1 5.570      1
#> 81     0  26   1 5.525      1
#> 82     1  22   1 6.775      0
#> 83     1  29   1 6.775      0
#> 84     0  50   1 5.448      1
#> 85     0  23   0 3.422      1
#> 86     0  41   0 3.542      1
#> 87     0  27   0 3.301      1
#> 88     0  24   1 5.539      1
#> 89     0  48   1 5.355      1
#> 90     0  47   0 3.486      1
#> 91     1  26   1 6.645      1
#> 92     1  28   0 4.983      1
#> 93     0  38   0 3.502      1
#> 94     0  33   0 3.484      1
#> 95     0  29   0 3.302      1
#> 96     0  23   1 5.523      1
#> 97     0  28   0 3.513      1
#> 98     0  40   1 5.372      1
#> 99     1  48   1 6.775      0
#> 100    1  21   1 6.746      1