Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      1  45   0 5.041      1
#> 2      1  39   1 6.876      0
#> 3      1  29   1 6.876      0
#> 4      0  42   1 5.438      1
#> 5      1  48   0 4.783      1
#> 6      1  24   1 6.876      0
#> 7      0  39   1 5.491      1
#> 8      1  23   0 4.899      1
#> 9      0  49   1 5.508      1
#> 10     1  39   1 6.876      0
#> 11     1  29   1 6.876      0
#> 12     1  44   1 6.876      0
#> 13     1  47   1 6.876      0
#> 14     0  47   0 3.329      1
#> 15     0  27   1 5.635      1
#> 16     1  22   0 4.767      1
#> 17     1  44   1 6.876      0
#> 18     0  49   1 5.419      1
#> 19     1  30   1 6.876      0
#> 20     1  50   1 6.876      0
#> 21     1  24   1 6.876      0
#> 22     1  43   1 6.876      0
#> 23     0  27   1 5.533      1
#> 24     0  45   1 5.321      1
#> 25     0  39   1 5.342      1
#> 26     0  31   0 3.535      1
#> 27     0  39   1 5.465      1
#> 28     0  24   1 5.392      1
#> 29     0  40   1 5.484      1
#> 30     0  38   1 5.383      1
#> 31     0  39   1 5.448      1
#> 32     1  24   1 6.876      0
#> 33     0  37   0 3.469      1
#> 34     1  43   1 6.876      0
#> 35     0  22   1 5.497      1
#> 36     1  46   1 6.833      1
#> 37     0  23   1 5.496      1
#> 38     1  42   1 6.876      0
#> 39     1  26   1 6.876      0
#> 40     0  29   0 3.339      1
#> 41     0  41   1 5.376      1
#> 42     0  48   1 5.534      1
#> 43     0  23   1 5.341      1
#> 44     0  45   1 5.549      1
#> 45     1  31   1 6.876      0
#> 46     0  22   1 5.489      1
#> 47     1  41   1 6.870      1
#> 48     0  34   0 3.520      1
#> 49     0  46   1 5.495      1
#> 50     0  45   1 5.382      1
#> 51     0  47   1 5.389      1
#> 52     1  36   1 6.719      1
#> 53     0  40   0 3.357      1
#> 54     1  27   1 6.876      0
#> 55     0  39   1 5.384      1
#> 56     1  26   1 6.876      0
#> 57     1  36   1 6.876      0
#> 58     0  34   0 3.492      1
#> 59     0  33   0 3.423      1
#> 60     1  38   1 6.852      1
#> 61     0  46   1 5.496      1
#> 62     0  40   0 3.445      1
#> 63     1  35   1 6.876      0
#> 64     0  26   1 4.869      1
#> 65     1  40   1 6.876      0
#> 66     0  37   1 5.557      1
#> 67     1  28   0 5.229      1
#> 68     1  36   0 5.010      1
#> 69     1  50   1 6.876      0
#> 70     0  35   1 5.449      1
#> 71     0  37   1 5.418      1
#> 72     0  35   1 5.429      1
#> 73     1  39   1 6.876      0
#> 74     0  44   0 3.425      1
#> 75     0  47   1 5.174      1
#> 76     0  38   1 5.563      1
#> 77     1  49   1 6.876      0
#> 78     0  41   1 5.525      1
#> 79     1  43   1 6.746      1
#> 80     1  24   1 6.876      0
#> 81     1  37   1 6.876      0
#> 82     0  40   1 5.509      1
#> 83     0  41   0 3.420      1
#> 84     1  26   1 6.876      0
#> 85     0  22   1 5.613      1
#> 86     1  24   1 6.876      0
#> 87     1  25   0 5.078      1
#> 88     0  34   1 5.452      1
#> 89     0  27   1 5.274      1
#> 90     0  20   1 5.505      1
#> 91     0  22   1 5.643      1
#> 92     1  33   1 6.838      1
#> 93     1  42   0 5.004      1
#> 94     0  42   1 5.397      1
#> 95     0  47   1 5.466      1
#> 96     0  39   0 3.468      1
#> 97     1  33   0 5.004      1
#> 98     1  34   1 6.876      0
#> 99     0  28   0 3.449      1
#> 100    0  40   1 5.558      1