Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      0  26   0 3.375      1
#> 2      1  20   1 6.764      0
#> 3      1  31   1 6.764      0
#> 4      1  34   1 6.764      0
#> 5      1  30   1 6.764      0
#> 6      1  22   1 6.764      0
#> 7      1  37   1 6.764      0
#> 8      0  31   1 5.367      1
#> 9      0  22   1 5.324      1
#> 10     0  35   0 3.528      1
#> 11     1  25   1 6.764      0
#> 12     0  38   1 5.478      1
#> 13     1  22   0 4.877      1
#> 14     1  29   1 6.552      1
#> 15     0  25   1 5.584      1
#> 16     0  39   1 5.548      1
#> 17     0  37   0 3.470      1
#> 18     1  28   1 6.764      0
#> 19     0  41   1 5.480      1
#> 20     1  23   1 6.764      0
#> 21     0  22   1 5.570      1
#> 22     0  36   1 5.571      1
#> 23     1  24   0 4.924      1
#> 24     0  36   1 5.406      1
#> 25     0  37   1 5.261      1
#> 26     0  33   1 5.563      1
#> 27     0  33   0 3.445      1
#> 28     1  21   1 6.764      0
#> 29     1  44   1 6.764      0
#> 30     1  41   1 6.764      0
#> 31     0  41   1 5.394      1
#> 32     1  49   0 4.934      1
#> 33     1  43   0 4.982      1
#> 34     1  41   1 6.764      0
#> 35     1  23   1 6.762      1
#> 36     0  36   1 5.469      1
#> 37     1  32   1 6.764      0
#> 38     0  45   1 5.500      1
#> 39     1  47   0 5.048      1
#> 40     0  46   1 5.513      1
#> 41     1  40   1 6.764      0
#> 42     1  50   1 6.764      0
#> 43     0  44   0 3.484      1
#> 44     0  50   0 3.428      1
#> 45     1  24   0 5.084      1
#> 46     0  39   1 5.562      1
#> 47     0  30   0 3.291      1
#> 48     1  45   1 6.764      0
#> 49     1  44   0 4.988      1
#> 50     0  25   1 5.623      1
#> 51     1  39   1 6.764      0
#> 52     1  38   1 6.764      0
#> 53     0  21   1 5.379      1
#> 54     0  22   1 5.614      1
#> 55     0  33   1 5.497      1
#> 56     1  25   1 6.764      0
#> 57     1  25   0 4.933      1
#> 58     1  32   0 4.893      1
#> 59     0  43   1 5.544      1
#> 60     1  33   0 4.999      1
#> 61     0  44   1 5.550      1
#> 62     0  48   1 5.456      1
#> 63     0  33   1 5.348      1
#> 64     0  38   0 3.495      1
#> 65     1  29   0 5.000      1
#> 66     1  37   1 6.764      0
#> 67     0  46   0 3.528      1
#> 68     0  24   1 5.482      1
#> 69     0  34   1 5.393      1
#> 70     0  26   1 5.550      1
#> 71     0  40   1 5.501      1
#> 72     1  48   0 5.002      1
#> 73     0  22   1 5.446      1
#> 74     0  50   1 5.510      1
#> 75     0  44   1 5.546      1
#> 76     0  48   1 5.431      1
#> 77     0  33   1 5.437      1
#> 78     1  26   0 4.736      1
#> 79     0  26   1 5.432      1
#> 80     1  30   0 4.787      1
#> 81     1  30   0 5.070      1
#> 82     1  43   1 6.764      0
#> 83     0  49   1 5.506      1
#> 84     1  43   1 6.764      0
#> 85     1  37   1 6.764      0
#> 86     1  24   1 6.764      0
#> 87     0  21   1 5.183      1
#> 88     0  22   1 5.489      1
#> 89     0  39   1 5.344      1
#> 90     0  24   1 5.685      1
#> 91     1  38   0 5.011      1
#> 92     0  50   1 5.508      1
#> 93     1  42   1 6.764      0
#> 94     0  38   1 5.336      1
#> 95     1  49   1 6.764      0
#> 96     1  40   1 6.764      0
#> 97     0  28   1 5.494      1
#> 98     1  46   1 6.764      0
#> 99     1  49   1 6.764      0
#> 100    1  27   0 5.076      1