Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      1  27   1 6.803      1
#> 2      0  32   0 3.486      1
#> 3      0  27   1 5.452      1
#> 4      0  24   1 5.687      1
#> 5      0  32   1 5.412      1
#> 6      1  24   1 6.819      0
#> 7      0  41   1 5.548      1
#> 8      0  42   1 5.403      1
#> 9      0  39   1 5.523      1
#> 10     1  26   0 5.059      1
#> 11     1  49   1 6.819      0
#> 12     0  41   1 5.581      1
#> 13     0  35   1 5.437      1
#> 14     0  49   1 5.470      1
#> 15     0  34   1 5.467      1
#> 16     1  30   1 6.812      1
#> 17     0  22   1 5.593      1
#> 18     0  24   0 3.502      1
#> 19     1  27   1 6.819      0
#> 20     0  30   0 3.564      1
#> 21     0  22   0 3.345      1
#> 22     1  38   1 6.819      0
#> 23     1  33   1 6.819      0
#> 24     1  27   1 6.819      0
#> 25     1  37   1 6.819      0
#> 26     1  31   1 6.819      0
#> 27     0  48   0 3.479      1
#> 28     0  29   1 5.471      1
#> 29     1  46   0 4.949      1
#> 30     0  43   1 5.406      1
#> 31     1  42   0 5.005      1
#> 32     1  26   1 6.819      0
#> 33     1  28   1 6.258      1
#> 34     0  31   1 5.465      1
#> 35     0  43   1 5.434      1
#> 36     0  30   0 3.499      1
#> 37     0  24   1 5.467      1
#> 38     1  34   1 6.776      1
#> 39     1  36   1 6.819      0
#> 40     0  34   0 3.345      1
#> 41     0  41   1 5.439      1
#> 42     1  38   1 6.819      0
#> 43     1  29   1 6.819      0
#> 44     1  23   0 5.155      1
#> 45     1  35   1 6.819      0
#> 46     0  30   1 5.467      1
#> 47     0  32   0 3.547      1
#> 48     1  36   0 4.992      1
#> 49     1  20   1 6.819      0
#> 50     0  29   0 3.457      1
#> 51     1  47   0 5.034      1
#> 52     1  30   1 6.819      0
#> 53     0  31   0 3.419      1
#> 54     0  37   0 3.480      1
#> 55     0  40   1 5.532      1
#> 56     1  32   0 4.947      1
#> 57     1  24   1 6.597      1
#> 58     1  48   1 6.819      0
#> 59     1  37   1 6.819      0
#> 60     1  33   0 5.055      1
#> 61     1  25   1 6.819      0
#> 62     0  34   1 5.568      1
#> 63     1  44   1 6.819      0
#> 64     1  42   0 4.963      1
#> 65     1  40   1 6.819      0
#> 66     0  33   0 3.506      1
#> 67     1  22   1 6.819      0
#> 68     1  40   0 4.970      1
#> 69     1  46   0 4.999      1
#> 70     0  21   0 3.500      1
#> 71     1  40   1 6.819      0
#> 72     0  31   0 3.490      1
#> 73     0  39   1 5.436      1
#> 74     1  21   1 6.807      1
#> 75     1  33   0 4.959      1
#> 76     1  42   1 6.819      0
#> 77     1  41   0 5.044      1
#> 78     1  20   1 6.767      1
#> 79     0  23   1 5.651      1
#> 80     0  36   1 5.506      1
#> 81     1  42   1 6.819      0
#> 82     0  22   1 5.522      1
#> 83     1  39   1 6.819      0
#> 84     0  33   1 5.566      1
#> 85     1  21   1 6.775      1
#> 86     1  23   1 6.819      0
#> 87     0  25   1 5.357      1
#> 88     1  38   1 6.819      0
#> 89     0  31   1 5.350      1
#> 90     1  40   1 6.819      0
#> 91     0  41   1 5.579      1
#> 92     1  48   0 5.022      1
#> 93     1  34   1 6.819      0
#> 94     0  29   1 5.470      1
#> 95     1  29   1 6.819      0
#> 96     1  38   0 4.909      1
#> 97     0  39   0 3.354      1
#> 98     0  35   1 5.566      1
#> 99     0  47   0 3.468      1
#> 100    1  24   0 5.014      1