Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      0  33   1 5.501      1
#> 2      1  44   1 6.750      1
#> 3      0  35   0 3.503      1
#> 4      0  24   1 5.459      1
#> 5      0  22   1 5.361      1
#> 6      1  50   0 4.922      1
#> 7      0  21   0 3.403      1
#> 8      0  22   0 3.550      1
#> 9      1  30   1 6.891      0
#> 10     1  23   1 6.891      0
#> 11     1  30   1 6.891      0
#> 12     0  46   0 3.423      1
#> 13     1  42   1 6.891      0
#> 14     0  35   1 5.555      1
#> 15     0  48   0 3.476      1
#> 16     1  34   1 6.787      1
#> 17     0  45   0 3.454      1
#> 18     1  35   0 4.962      1
#> 19     1  43   1 6.891      0
#> 20     0  44   1 5.433      1
#> 21     1  38   1 6.891      0
#> 22     1  47   0 4.957      1
#> 23     1  41   1 6.891      0
#> 24     1  37   1 6.891      0
#> 25     1  43   1 6.886      1
#> 26     1  27   0 4.793      1
#> 27     0  32   1 5.055      1
#> 28     1  41   1 6.891      0
#> 29     0  22   0 3.458      1
#> 30     0  35   1 5.582      1
#> 31     0  23   1 5.490      1
#> 32     0  49   0 3.513      1
#> 33     0  24   0 3.437      1
#> 34     0  44   1 5.444      1
#> 35     0  35   1 5.551      1
#> 36     1  43   1 6.891      0
#> 37     1  28   1 6.891      0
#> 38     0  46   1 5.326      1
#> 39     0  46   1 5.544      1
#> 40     0  28   1 5.537      1
#> 41     1  37   1 6.891      0
#> 42     1  41   1 6.891      0
#> 43     0  21   1 5.285      1
#> 44     1  35   1 6.891      0
#> 45     0  37   1 5.579      1
#> 46     1  43   0 4.959      1
#> 47     1  41   1 6.891      0
#> 48     0  49   1 5.484      1
#> 49     1  49   1 6.891      0
#> 50     1  45   0 5.003      1
#> 51     0  35   0 3.381      1
#> 52     1  26   1 6.579      1
#> 53     0  23   0 3.400      1
#> 54     0  26   0 3.338      1
#> 55     0  34   1 5.196      1
#> 56     1  39   1 6.891      0
#> 57     0  28   1 5.259      1
#> 58     1  46   1 6.739      1
#> 59     1  47   1 6.891      0
#> 60     0  39   0 3.362      1
#> 61     1  46   1 6.891      0
#> 62     1  31   1 6.891      0
#> 63     0  21   1 5.177      1
#> 64     1  32   1 6.651      1
#> 65     1  48   1 6.891      0
#> 66     0  41   1 5.495      1
#> 67     1  20   1 6.891      0
#> 68     1  23   1 6.891      0
#> 69     0  26   1 5.539      1
#> 70     0  31   1 5.529      1
#> 71     1  30   1 6.891      0
#> 72     0  29   1 5.157      1
#> 73     0  29   1 5.540      1
#> 74     0  32   0 3.561      1
#> 75     1  27   1 6.891      0
#> 76     0  45   0 3.433      1
#> 77     0  47   1 5.246      1
#> 78     0  49   1 5.522      1
#> 79     0  37   1 5.538      1
#> 80     1  40   1 6.891      0
#> 81     0  43   0 3.484      1
#> 82     1  25   1 6.891      0
#> 83     0  34   1 5.575      1
#> 84     0  25   1 5.339      1
#> 85     1  37   1 6.891      0
#> 86     0  33   1 5.633      1
#> 87     0  36   1 5.496      1
#> 88     0  37   1 5.468      1
#> 89     1  40   1 6.887      1
#> 90     1  33   1 6.891      0
#> 91     1  31   1 6.815      1
#> 92     0  32   0 3.477      1
#> 93     1  26   1 6.838      1
#> 94     0  22   1 5.581      1
#> 95     0  26   1 5.484      1
#> 96     0  38   1 5.454      1
#> 97     1  31   1 6.891      0
#> 98     0  39   0 3.453      1
#> 99     1  39   1 6.766      1
#> 100    0  22   0 3.570      1