Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      1  38   1 6.512      1
#> 2      1  32   1 6.796      0
#> 3      0  29   1 5.461      1
#> 4      0  49   0 3.520      1
#> 5      1  31   0 4.924      1
#> 6      1  40   1 6.796      0
#> 7      1  49   1 6.796      0
#> 8      1  35   0 5.040      1
#> 9      1  49   0 5.071      1
#> 10     0  32   0 3.445      1
#> 11     1  27   1 6.622      1
#> 12     0  21   0 3.445      1
#> 13     1  42   1 6.796      0
#> 14     1  24   0 5.041      1
#> 15     1  44   0 4.966      1
#> 16     1  35   0 4.949      1
#> 17     1  35   1 6.656      1
#> 18     0  21   1 5.516      1
#> 19     0  38   0 3.505      1
#> 20     1  40   1 6.796      0
#> 21     0  23   0 3.430      1
#> 22     0  38   1 5.472      1
#> 23     0  31   1 5.433      1
#> 24     0  48   1 5.577      1
#> 25     1  27   1 6.613      1
#> 26     1  46   0 5.047      1
#> 27     1  33   1 6.796      0
#> 28     0  43   1 5.179      1
#> 29     1  42   1 6.796      0
#> 30     0  29   0 3.468      1
#> 31     1  34   1 6.796      0
#> 32     1  32   1 6.796      0
#> 33     0  33   1 5.252      1
#> 34     0  33   1 5.585      1
#> 35     0  43   1 5.417      1
#> 36     1  32   1 6.796      0
#> 37     1  33   1 6.796      0
#> 38     0  34   1 5.524      1
#> 39     0  32   1 5.361      1
#> 40     1  41   1 6.796      0
#> 41     0  28   0 3.507      1
#> 42     0  48   0 3.438      1
#> 43     1  34   1 6.796      0
#> 44     1  26   1 6.705      1
#> 45     1  46   0 5.064      1
#> 46     0  32   1 5.269      1
#> 47     0  25   0 3.540      1
#> 48     0  26   1 5.380      1
#> 49     0  38   0 3.519      1
#> 50     1  34   1 6.796      0
#> 51     1  29   1 6.796      0
#> 52     1  48   1 6.796      0
#> 53     1  20   1 6.796      0
#> 54     0  29   1 5.114      1
#> 55     0  33   1 5.388      1
#> 56     1  24   1 6.586      1
#> 57     1  47   0 5.014      1
#> 58     1  48   0 4.982      1
#> 59     0  23   1 5.546      1
#> 60     1  29   1 6.796      0
#> 61     1  49   0 4.962      1
#> 62     0  30   1 5.587      1
#> 63     1  39   1 6.796      0
#> 64     1  31   1 6.785      1
#> 65     1  35   1 6.796      0
#> 66     1  27   0 4.968      1
#> 67     1  36   1 6.796      0
#> 68     0  34   0 3.555      1
#> 69     0  20   1 5.189      1
#> 70     1  23   1 6.796      0
#> 71     1  48   1 6.796      0
#> 72     0  46   1 5.513      1
#> 73     1  27   0 4.716      1
#> 74     0  37   1 5.345      1
#> 75     0  38   1 5.464      1
#> 76     1  30   1 6.796      0
#> 77     0  35   0 3.458      1
#> 78     1  46   1 6.796      0
#> 79     0  29   0 3.380      1
#> 80     1  46   1 6.796      0
#> 81     0  22   1 5.229      1
#> 82     1  46   0 5.010      1
#> 83     1  44   1 6.796      0
#> 84     1  42   1 6.793      1
#> 85     0  37   1 5.492      1
#> 86     0  20   1 5.119      1
#> 87     0  42   1 5.507      1
#> 88     1  47   0 4.945      1
#> 89     1  40   1 6.796      0
#> 90     1  44   0 4.983      1
#> 91     0  40   1 5.327      1
#> 92     1  31   0 4.986      1
#> 93     0  42   1 5.519      1
#> 94     0  35   0 3.466      1
#> 95     1  40   0 4.790      1
#> 96     0  44   0 3.491      1
#> 97     1  34   1 6.796      0
#> 98     0  25   1 5.389      1
#> 99     0  28   1 5.578      1
#> 100    1  22   1 6.796      0