Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      1  33   1 6.742      0
#> 2      1  23   1 6.742      0
#> 3      0  20   0 3.328      1
#> 4      0  30   1 5.612      1
#> 5      0  42   1 5.439      1
#> 6      1  35   1 6.742      0
#> 7      0  46   1 5.446      1
#> 8      0  21   1 5.480      1
#> 9      1  50   1 6.742      0
#> 10     0  30   1 5.469      1
#> 11     1  38   1 6.742      0
#> 12     0  21   0 3.465      1
#> 13     1  35   0 5.047      1
#> 14     1  37   0 4.906      1
#> 15     1  45   0 5.046      1
#> 16     0  37   0 3.440      1
#> 17     0  22   1 5.326      1
#> 18     1  30   1 6.742      0
#> 19     1  50   1 6.742      0
#> 20     1  31   1 6.742      0
#> 21     1  25   1 6.742      0
#> 22     0  26   1 5.487      1
#> 23     1  42   0 5.021      1
#> 24     1  22   1 6.742      0
#> 25     1  47   1 6.742      0
#> 26     0  39   1 5.510      1
#> 27     1  37   1 6.742      0
#> 28     0  37   1 5.314      1
#> 29     1  47   1 6.742      0
#> 30     1  44   0 4.959      1
#> 31     1  43   1 6.742      0
#> 32     1  47   0 4.965      1
#> 33     0  22   1 5.620      1
#> 34     1  23   1 6.742      0
#> 35     0  37   1 5.452      1
#> 36     1  37   1 6.742      0
#> 37     1  30   1 6.742      0
#> 38     1  41   1 6.742      0
#> 39     1  41   0 5.073      1
#> 40     0  33   1 5.372      1
#> 41     0  35   0 3.314      1
#> 42     1  48   1 6.742      0
#> 43     1  41   1 6.742      0
#> 44     1  35   1 6.682      1
#> 45     0  36   0 3.523      1
#> 46     1  24   0 4.971      1
#> 47     0  25   1 5.504      1
#> 48     1  45   1 6.742      0
#> 49     0  31   0 3.504      1
#> 50     1  43   1 6.742      0
#> 51     0  38   1 5.333      1
#> 52     1  45   0 4.948      1
#> 53     0  25   1 5.558      1
#> 54     1  33   0 4.994      1
#> 55     0  29   0 3.373      1
#> 56     0  40   1 5.484      1
#> 57     0  33   1 5.518      1
#> 58     0  35   0 3.512      1
#> 59     1  34   0 4.946      1
#> 60     1  35   1 6.742      0
#> 61     0  35   1 5.555      1
#> 62     0  45   0 3.513      1
#> 63     0  43   1 5.414      1
#> 64     1  46   1 6.742      0
#> 65     0  49   0 3.471      1
#> 66     0  34   1 5.330      1
#> 67     1  39   1 6.742      0
#> 68     0  21   1 5.105      1
#> 69     0  28   1 5.432      1
#> 70     0  47   1 5.554      1
#> 71     0  41   1 5.614      1
#> 72     0  47   1 5.500      1
#> 73     1  39   0 4.963      1
#> 74     0  35   1 5.394      1
#> 75     0  49   0 3.539      1
#> 76     1  23   0 4.681      1
#> 77     1  49   1 6.742      0
#> 78     0  20   1 5.440      1
#> 79     1  30   1 6.742      0
#> 80     0  29   1 5.539      1
#> 81     0  41   1 5.522      1
#> 82     0  41   1 5.559      1
#> 83     0  22   1 5.233      1
#> 84     0  39   1 5.526      1
#> 85     1  36   1 6.742      0
#> 86     0  39   1 5.309      1
#> 87     0  25   0 3.493      1
#> 88     1  27   0 4.922      1
#> 89     0  49   1 5.442      1
#> 90     0  42   1 5.504      1
#> 91     1  41   0 4.983      1
#> 92     0  21   1 5.474      1
#> 93     1  41   0 5.042      1
#> 94     1  37   1 6.742      0
#> 95     0  44   0 3.441      1
#> 96     0  49   0 3.410      1
#> 97     1  42   0 5.003      1
#> 98     1  33   1 6.742      0
#> 99     0  34   0 3.456      1
#> 100    1  25   1 6.732      1