Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      1  49   0 5.032      1
#> 2      1  27   1 6.718      0
#> 3      0  44   1 5.446      1
#> 4      1  32   1 6.718      0
#> 5      1  34   1 6.718      0
#> 6      0  25   1 5.106      1
#> 7      1  32   0 4.938      1
#> 8      1  33   1 6.694      1
#> 9      0  44   1 5.425      1
#> 10     1  33   0 4.851      1
#> 11     1  41   0 5.100      1
#> 12     0  28   1 5.334      1
#> 13     1  45   0 5.072      1
#> 14     1  32   1 6.718      0
#> 15     1  46   1 6.718      0
#> 16     0  43   1 5.462      1
#> 17     1  34   1 6.718      0
#> 18     0  50   1 5.479      1
#> 19     0  29   1 5.360      1
#> 20     1  47   1 6.718      0
#> 21     1  39   1 6.718      0
#> 22     1  33   1 6.718      0
#> 23     0  30   1 5.528      1
#> 24     1  33   1 6.718      0
#> 25     0  48   1 5.456      1
#> 26     1  26   0 4.956      1
#> 27     1  43   1 6.718      0
#> 28     0  41   1 5.586      1
#> 29     0  22   1 5.628      1
#> 30     1  32   1 6.718      0
#> 31     0  37   1 5.463      1
#> 32     1  42   0 5.040      1
#> 33     1  31   1 6.718      0
#> 34     1  42   0 4.952      1
#> 35     1  44   1 6.718      0
#> 36     1  23   0 4.960      1
#> 37     0  26   1 5.274      1
#> 38     0  40   1 5.430      1
#> 39     0  43   1 5.455      1
#> 40     1  29   1 6.598      1
#> 41     0  24   1 5.414      1
#> 42     1  26   0 5.055      1
#> 43     0  28   1 5.423      1
#> 44     0  42   1 5.510      1
#> 45     0  27   1 5.348      1
#> 46     0  41   1 5.529      1
#> 47     1  30   0 5.077      1
#> 48     1  45   1 6.718      0
#> 49     0  42   0 3.455      1
#> 50     1  32   0 5.033      1
#> 51     0  47   1 5.565      1
#> 52     0  42   1 5.471      1
#> 53     0  32   1 5.474      1
#> 54     0  48   1 5.438      1
#> 55     0  41   1 5.551      1
#> 56     0  26   0 3.558      1
#> 57     0  41   1 5.532      1
#> 58     0  44   1 5.315      1
#> 59     1  43   1 6.718      0
#> 60     0  33   1 5.563      1
#> 61     1  39   1 6.718      0
#> 62     0  27   1 5.264      1
#> 63     1  29   0 4.998      1
#> 64     0  39   1 5.407      1
#> 65     1  29   0 4.940      1
#> 66     1  43   1 6.718      0
#> 67     1  44   1 6.718      0
#> 68     1  48   0 5.050      1
#> 69     0  28   0 3.505      1
#> 70     1  33   1 6.718      0
#> 71     0  20   1 5.485      1
#> 72     0  39   1 5.479      1
#> 73     1  21   1 6.718      0
#> 74     1  28   1 6.718      0
#> 75     0  36   0 3.477      1
#> 76     1  29   1 6.718      0
#> 77     0  48   1 5.356      1
#> 78     1  44   1 6.718      0
#> 79     0  41   1 5.465      1
#> 80     1  47   1 6.718      0
#> 81     1  26   1 6.718      0
#> 82     0  35   1 5.458      1
#> 83     1  23   0 4.891      1
#> 84     1  38   1 6.718      0
#> 85     0  21   0 3.555      1
#> 86     1  41   0 4.861      1
#> 87     1  28   1 6.718      0
#> 88     1  39   0 4.953      1
#> 89     1  31   0 5.098      1
#> 90     0  33   1 5.361      1
#> 91     1  29   0 4.917      1
#> 92     0  34   1 5.466      1
#> 93     1  26   1 6.718      0
#> 94     1  44   1 6.412      1
#> 95     1  46   1 6.718      0
#> 96     1  27   0 4.911      1
#> 97     0  25   0 3.479      1
#> 98     0  39   0 3.441      1
#> 99     0  39   0 3.414      1
#> 100    0  41   1 5.547      1