Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cutoff = NULL, cens = 0.3)

Arguments

n
(integer(1))
Number of samples

trt, age, sex
(numeric(1))
Coefficients for covariates.

cutoff
(numeric(1))
Deprecated, in future use cens.

cens
(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()

#>     sexF age trt  time status
#> 1      1  36   0 5.079      1
#> 2      0  40   1 5.394      1
#> 3      1  49   0 4.961      1
#> 4      1  37   1 6.900      0
#> 5      1  43   1 6.900      0
#> 6      0  23   1 5.449      1
#> 7      0  35   1 5.361      1
#> 8      1  46   0 4.920      1
#> 9      1  45   1 6.900      0
#> 10     0  38   1 5.496      1
#> 11     1  36   1 6.900      0
#> 12     1  41   0 5.026      1
#> 13     1  49   1 6.900      0
#> 14     0  29   1 5.465      1
#> 15     1  42   1 6.571      1
#> 16     0  47   1 5.545      1
#> 17     0  28   0 3.299      1
#> 18     0  23   1 5.404      1
#> 19     1  34   0 5.059      1
#> 20     1  42   1 6.900      0
#> 21     1  27   0 4.973      1
#> 22     0  45   1 5.423      1
#> 23     0  23   1 5.586      1
#> 24     1  31   0 4.976      1
#> 25     1  41   1 6.875      1
#> 26     0  34   1 5.562      1
#> 27     1  28   0 4.900      1
#> 28     1  21   0 4.973      1
#> 29     1  40   0 5.020      1
#> 30     1  25   1 6.900      0
#> 31     1  44   0 4.975      1
#> 32     0  44   1 5.568      1
#> 33     1  35   1 6.900      0
#> 34     1  41   1 6.900      0
#> 35     1  26   1 6.770      1
#> 36     1  45   1 6.825      1
#> 37     1  47   0 5.047      1
#> 38     1  44   0 5.062      1
#> 39     1  39   1 6.900      0
#> 40     0  40   1 5.556      1
#> 41     1  27   1 6.900      0
#> 42     1  28   0 4.753      1
#> 43     0  30   1 5.558      1
#> 44     0  41   1 5.355      1
#> 45     0  20   0 3.560      1
#> 46     1  29   1 6.900      0
#> 47     1  25   1 6.900      0
#> 48     0  33   1 5.525      1
#> 49     1  47   1 6.900      0
#> 50     0  27   1 5.413      1
#> 51     0  36   1 5.563      1
#> 52     0  28   1 5.273      1
#> 53     0  22   1 5.593      1
#> 54     0  49   1 5.482      1
#> 55     1  36   1 6.900      0
#> 56     1  44   1 6.900      0
#> 57     1  20   0 5.021      1
#> 58     1  39   1 6.721      1
#> 59     1  39   1 6.900      0
#> 60     0  30   0 3.499      1
#> 61     0  48   0 3.436      1
#> 62     1  44   1 6.900      0
#> 63     0  49   1 5.376      1
#> 64     0  46   1 5.447      1
#> 65     1  38   1 6.894      1
#> 66     1  34   1 6.900      0
#> 67     0  27   1 5.512      1
#> 68     0  44   1 5.422      1
#> 69     1  24   1 6.852      1
#> 70     1  46   1 6.900      0
#> 71     1  45   0 4.956      1
#> 72     1  32   1 6.726      1
#> 73     1  44   1 6.900      0
#> 74     0  31   1 5.316      1
#> 75     1  21   1 6.523      1
#> 76     1  32   1 6.900      0
#> 77     0  38   1 5.535      1
#> 78     1  26   1 6.900      0
#> 79     0  21   0 3.602      1
#> 80     1  34   1 6.753      1
#> 81     0  36   0 3.500      1
#> 82     1  34   0 5.026      1
#> 83     0  38   0 3.566      1
#> 84     0  43   1 5.284      1
#> 85     1  28   1 6.900      0
#> 86     1  36   1 6.685      1
#> 87     1  34   1 6.900      0
#> 88     1  47   1 6.900      0
#> 89     1  43   1 6.900      0
#> 90     0  24   0 3.498      1
#> 91     1  33   1 6.890      1
#> 92     1  26   1 6.900      0
#> 93     1  27   1 6.900      0
#> 94     1  22   0 5.031      1
#> 95     1  28   1 6.778      1
#> 96     1  32   1 6.900      0
#> 97     0  23   1 5.642      1
#> 98     0  36   0 3.372      1
#> 99     1  30   0 5.064      1
#> 100    0  35   0 3.459      1