Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      1  25   0 4.959      1
#> 2      0  24   1 5.420      1
#> 3      1  28   1 6.685      0
#> 4      1  44   1 6.685      0
#> 5      0  23   0 3.546      1
#> 6      1  37   0 4.906      1
#> 7      0  45   1 5.560      1
#> 8      1  28   0 5.010      1
#> 9      1  46   1 6.685      0
#> 10     0  36   1 5.297      1
#> 11     1  26   1 6.685      0
#> 12     1  34   1 6.685      0
#> 13     1  24   0 4.961      1
#> 14     1  37   0 5.038      1
#> 15     1  25   1 6.685      0
#> 16     1  34   1 6.685      0
#> 17     0  23   0 3.365      1
#> 18     1  43   0 4.898      1
#> 19     0  37   1 5.461      1
#> 20     1  20   0 4.428      1
#> 21     1  29   1 6.685      0
#> 22     0  25   1 5.544      1
#> 23     0  35   1 5.529      1
#> 24     0  26   1 5.278      1
#> 25     1  43   0 4.928      1
#> 26     1  37   1 6.685      0
#> 27     1  26   0 5.024      1
#> 28     0  46   0 3.329      1
#> 29     1  26   0 4.931      1
#> 30     1  32   0 4.975      1
#> 31     0  41   1 5.492      1
#> 32     0  39   1 5.492      1
#> 33     0  33   1 5.533      1
#> 34     0  28   1 5.460      1
#> 35     1  43   0 4.971      1
#> 36     1  21   1 6.685      0
#> 37     1  27   1 6.685      0
#> 38     0  36   1 5.525      1
#> 39     1  45   0 4.940      1
#> 40     1  36   1 6.685      0
#> 41     0  41   1 5.499      1
#> 42     0  31   1 5.457      1
#> 43     1  26   1 6.685      0
#> 44     1  47   1 6.685      0
#> 45     1  46   0 5.021      1
#> 46     1  44   0 4.951      1
#> 47     0  27   1 5.440      1
#> 48     1  26   1 6.685      0
#> 49     0  32   0 3.461      1
#> 50     1  44   1 6.685      0
#> 51     1  26   1 6.685      0
#> 52     0  21   1 5.381      1
#> 53     0  49   1 5.448      1
#> 54     1  21   0 4.979      1
#> 55     0  39   1 5.497      1
#> 56     1  25   1 6.656      1
#> 57     0  43   1 5.550      1
#> 58     1  38   0 4.943      1
#> 59     1  30   1 6.685      0
#> 60     0  49   1 5.381      1
#> 61     1  21   1 6.685      0
#> 62     0  41   1 5.411      1
#> 63     0  48   1 5.535      1
#> 64     1  35   1 6.685      0
#> 65     1  28   1 6.685      0
#> 66     0  21   0 3.537      1
#> 67     0  21   1 5.645      1
#> 68     1  21   1 6.685      0
#> 69     1  37   1 6.685      0
#> 70     1  20   1 6.685      0
#> 71     0  26   1 5.295      1
#> 72     0  30   1 5.616      1
#> 73     1  50   0 4.935      1
#> 74     1  21   1 6.685      0
#> 75     0  25   1 5.599      1
#> 76     1  22   1 6.330      1
#> 77     0  20   1 5.167      1
#> 78     1  37   0 5.063      1
#> 79     1  21   1 6.685      0
#> 80     0  30   0 3.488      1
#> 81     0  37   1 5.532      1
#> 82     0  38   0 3.506      1
#> 83     1  46   0 5.008      1
#> 84     0  43   1 5.475      1
#> 85     1  37   1 6.685      0
#> 86     1  34   1 6.685      0
#> 87     1  37   0 4.966      1
#> 88     1  29   0 4.840      1
#> 89     1  23   0 4.982      1
#> 90     0  37   1 5.550      1
#> 91     0  50   0 3.358      1
#> 92     0  39   1 5.291      1
#> 93     1  38   1 6.685      0
#> 94     0  31   0 3.477      1
#> 95     0  36   0 3.449      1
#> 96     0  34   1 5.521      1
#> 97     0  37   0 3.509      1
#> 98     1  46   1 6.685      0
#> 99     1  27   0 4.989      1
#> 100    0  43   1 5.369      1