Simulate Survival Data

Function for simulating survival data.

simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cens = 0.3)

Arguments

n

(integer(1))
Number of samples

trt, age, sex

(numeric(1))
Coefficients for covariates.

cens

(numeric(1))
Proportion of censoring to be generated, cut-off time is then selected as the quantile that results in cens.

Value

data.frame()

Details

Currently limited to three covariates, Weibull survival times, and Type I censoring. This will be expanded to a flexible simulation function in future updates. For now the function is primarily limited to helping function examples.

Examples

simsurvdata()
#>     sexF age trt  time status
#> 1      0  22   0 3.307      1
#> 2      0  41   1 5.345      1
#> 3      1  50   0 4.909      1
#> 4      1  47   1 6.846      0
#> 5      0  20   1 5.235      1
#> 6      1  28   1 6.846      0
#> 7      1  28   1 6.846      0
#> 8      0  32   0 3.545      1
#> 9      0  27   1 5.510      1
#> 10     0  27   1 5.439      1
#> 11     0  31   0 3.367      1
#> 12     1  31   0 5.018      1
#> 13     0  31   1 5.465      1
#> 14     0  25   1 5.370      1
#> 15     0  24   0 3.420      1
#> 16     0  23   1 5.188      1
#> 17     1  43   1 6.846      0
#> 18     0  38   1 5.450      1
#> 19     1  30   1 6.846      0
#> 20     0  47   1 5.425      1
#> 21     0  48   1 5.470      1
#> 22     0  32   1 5.602      1
#> 23     1  46   0 4.934      1
#> 24     1  21   1 6.802      1
#> 25     0  30   0 3.403      1
#> 26     1  23   1 6.846      0
#> 27     0  32   0 3.483      1
#> 28     0  28   1 5.129      1
#> 29     1  37   1 6.846      0
#> 30     1  34   1 6.846      0
#> 31     1  30   1 6.846      0
#> 32     0  50   1 5.393      1
#> 33     1  30   1 6.846      0
#> 34     1  44   0 5.045      1
#> 35     1  28   1 6.846      0
#> 36     1  41   1 6.846      0
#> 37     0  23   0 3.444      1
#> 38     0  34   1 5.461      1
#> 39     1  31   1 6.846      0
#> 40     1  25   1 6.846      0
#> 41     0  21   1 5.175      1
#> 42     1  26   1 6.846      0
#> 43     0  41   1 5.339      1
#> 44     0  29   0 3.475      1
#> 45     1  46   1 6.846      0
#> 46     0  46   1 5.230      1
#> 47     0  30   1 5.494      1
#> 48     0  40   1 5.437      1
#> 49     1  42   1 6.846      0
#> 50     1  48   1 6.846      0
#> 51     1  35   0 4.816      1
#> 52     1  24   1 6.846      0
#> 53     1  48   1 6.846      0
#> 54     0  23   1 5.701      1
#> 55     0  34   0 3.416      1
#> 56     0  41   0 3.527      1
#> 57     0  28   0 3.389      1
#> 58     1  29   0 4.864      1
#> 59     1  31   1 6.846      0
#> 60     1  48   1 6.846      0
#> 61     1  47   1 6.846      0
#> 62     1  47   1 6.846      0
#> 63     0  30   1 5.600      1
#> 64     1  48   1 6.846      0
#> 65     1  41   1 6.846      0
#> 66     1  32   0 4.970      1
#> 67     0  30   1 5.591      1
#> 68     0  31   0 3.575      1
#> 69     1  25   1 6.846      0
#> 70     0  46   1 5.582      1
#> 71     0  47   1 5.475      1
#> 72     1  39   0 4.851      1
#> 73     0  28   0 3.483      1
#> 74     0  26   1 5.463      1
#> 75     0  45   1 5.555      1
#> 76     1  27   1 6.846      0
#> 77     0  35   1 5.566      1
#> 78     0  48   0 3.427      1
#> 79     1  44   1 6.750      1
#> 80     0  22   1 5.643      1
#> 81     1  28   0 4.776      1
#> 82     0  21   1 5.563      1
#> 83     0  39   1 5.516      1
#> 84     0  43   1 5.475      1
#> 85     1  47   0 4.845      1
#> 86     0  50   1 5.402      1
#> 87     0  32   0 3.268      1
#> 88     1  25   0 4.899      1
#> 89     0  36   1 5.470      1
#> 90     0  46   0 3.416      1
#> 91     1  44   1 6.846      0
#> 92     1  35   0 4.934      1
#> 93     0  44   1 5.529      1
#> 94     1  20   0 4.908      1
#> 95     0  24   1 5.381      1
#> 96     1  32   0 5.084      1
#> 97     1  40   0 4.935      1
#> 98     0  33   0 3.617      1
#> 99     1  33   1 6.836      1
#> 100    1  41   1 6.846      0